### 2025-07-01 #8530 @editorialbot
[PRE REVIEW]: rdkit2ase: Molecular Structure Generation and Manipulation for Machine-Learned Interatomic Potentials
**Submitting author:** @PythonFZ (<a href="http://orcid.org/0000-0002-6936-4692">Fabian Zills</a>)
**Repository:** https://github.com/zincware/rdkit2ase
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.1.8
