### 2025-07-07 56cdf4d
add tests for shifted example system
