### 2025-07-07 c010390
update test case
