+++ tests/test_examples.py
@@ -13,12 +13,13 @@ EMC= "CCOC(=O)OC"
 VC = "C1=COC(=O)O1"
 DMC = "COC(=O)OC"
 
-def test_example():
+def test_partial_suggestions():
     atoms = ase.io.read(DATA / "ec_emc.xyz")
     wrapped_atoms = atoms.copy()
     wrapped_atoms.positions += [5, 5, 5]
     wrapped_atoms.wrap()
 
+    # it will work, if all suggestions are present
     a = rdkit2ase.ase2networkx(
         atoms, suggestions=[VC, DMC]
     )
