### 2025-07-07 f0cad28
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
