+++ rdkit2ase/bond_order.py
@@ -1,11 +1,7 @@
 import networkx as nx
 from networkx.algorithms import isomorphism
 
-from rdkit2ase.utils import rdkit_determine_bonds, unwrap_molecule
-from rdkit2ase.utils import suggestions2networkx
-
-
-
+from rdkit2ase.utils import rdkit_determine_bonds, suggestions2networkx, unwrap_molecule
 
 
 def sort_templates(graphs: list[nx.Graph]) -> list[nx.Graph]:
