### 2025-07-07 6e69bcf
Add test for partial suggestions (#52)
* add tests for shifted example system

* update test

* update test case

* bugfix

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* update the subgraph

* remove print statement

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
