### 2025-07-14 60e219b
add vesin and `pbc` flag
