### 2025-07-14 a2b3a1b
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
