+++ pyproject.toml
@@ -12,6 +12,7 @@ dependencies = [
     "ase>=3.25.0",
     "networkx>=3.4.2",
     "rdkit>=2024",
+
 ]
 
 [dependency-groups]
@@ -27,6 +28,12 @@ docs = [
     "nbsphinx>=0.9.7",
     "sphinx>=8.1.3",
     "sphinx-copybutton>=0.5.2",
+]
+
+[project.optional-dependencies]
+
+vesin = [
+    "vesin>=0.3.7",
 ]
 
 [tool.ruff.lint]
