### 2025-07-15 #8530 @lucydot
Hi @PythonFZ I'll be your editor for this submission.

If you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
