### 2025-07-16 f0630a0
empty molecule edge cases
