+++ rdkit2ase/__init__.py
@@ -5,6 +5,7 @@ from rdkit2ase.pack import pack
 from rdkit2ase.rdkit2x import rdkit2ase, rdkit2networkx
 from rdkit2ase.smiles2x import smiles2atoms, smiles2conformers
 from rdkit2ase.substructure import get_substructures, iter_fragments, match_substructure
+from rdkit2ase.utils import unwrap_structures
 
 __all__ = [
     "ase2rdkit",
@@ -15,6 +16,7 @@ __all__ = [
     "match_substructure",
     "get_substructures",
     "iter_fragments",
+    "unwrap_structures",
     #
     "ase2networkx",
     #
