+++ rdkit2ase/utils.py
@@ -82,7 +82,69 @@ def calculate_box_dimensions(images: list[ase.Atoms], density: float) -> list[fl
 
 
 def unwrap_structures(atoms, scale=1.2) -> ase.Atoms:
-    """Robust unwrapping across PBC using root-referenced traversal."""
+    """Unwrap molecular structures across periodic boundary conditions (PBC).
+    
+    This function corrects atomic positions that have been wrapped across periodic 
+    boundaries, ensuring that bonded atoms appear as continuous molecular structures.
+    It can handle multiple disconnected molecules within the same unit cell.
+    
+    The algorithm works by:
+    1. Building a connectivity graph based on covalent radii
+    2. Traversing each connected component (molecule) using depth-first search
+    3. Accumulating periodic image shifts to maintain molecular connectivity
+    4. Applying the shifts to obtain unwrapped coordinates
+    
+    Parameters
+    ----------
+    atoms : ase.Atoms
+        The ASE Atoms object containing wrapped atomic positions
+    scale : float, optional
+        Scale factor for covalent radii cutoffs used in bond detection.
+        Larger values include more distant neighbors as bonded.
+        Default is 1.2.
+        
+    Returns
+    -------
+    ase.Atoms
+        A new ASE Atoms object with unwrapped atomic positions. The original
+        atoms object is not modified.
+        
+    Notes
+    -----
+    - The function preserves the original atoms object and returns a copy
+    - Works with any periodic boundary conditions (1D, 2D, or 3D)
+    - Handles multiple disconnected molecules/fragments
+    - Uses covalent radii scaled by the `scale` parameter for bond detection
+    
+    Examples
+    --------
+    >>> import numpy as np
+    >>> from rdkit2ase import smiles2conformers, pack, unwrap_structures
+    >>> 
+    >>> # Create a realistic molecular system
+    >>> water = smiles2conformers("O", numConfs=1)
+    >>> ethanol = smiles2conformers("CCO", numConfs=1)
+    >>> 
+    >>> # Pack molecules into a periodic box
+    >>> packed_system = pack(
+    ...     data=[water, ethanol], 
+    ...     counts=[10, 5], 
+    ...     density=800
+    ... )
+    >>> 
+    >>> # Simulate wrapped coordinates (as might occur in MD)
+    >>> cell = packed_system.get_cell()
+    >>> positions = packed_system.get_positions()
+    >>> 
+    >>> # Artificially move the box and wrap it again
+    >>> positions += np.array([cell[0, 0] * 0.5, 0, 0])  # Shift by half box length
+    >>> wrapped_atoms = packed_system.copy()
+    >>> wrapped_atoms.set_positions(positions)
+    >>> wrapped_atoms.wrap()  # Wrap back into PBC
+    >>> 
+    >>> # Unwrap the structures to get continuous molecules
+    >>> unwrapped_atoms = unwrap_structures(wrapped_atoms)
+    """
     positions = atoms.get_positions()
     cell = atoms.get_cell()
     numbers = atoms.get_atomic_numbers()
