### 2025-07-16 5bf19a9
add `scale` parameter
