+++ tests/test_utils.py
@@ -81,59 +81,6 @@ def make_molecule(shift=(0, 0, 0), cell=(10, 10, 10), pbc=True):
     return atoms
 
 
-@pytest.mark.parametrize(
-    "shift,expected",
-    [
-        ((0, 0, 0), 1.1),  # Normal
-        ((9.5, 0, 0), 1.1),  # Cross x boundary
-        ((0, 9.5, 0), 1.1),  # Cross y boundary
-        ((0, 0, 9.5), 1.1),  # Cross z boundary
-        ((9.5, 9.5, 9.5), 1.1),  # Cross all boundaries
-    ],
-)
-def test_unwrap_diatomic(shift, expected):
-    """Ensure diatomics get unwrapped correctly across boundaries."""
-    atoms = make_molecule(shift=shift)
-    atoms_wrapped = atoms.copy()
-    atoms_unwrapped = unwrap_structures(atoms_wrapped)
-
-    dist = atoms_unwrapped.get_distance(0, 1, mic=True)
-    assert np.isclose(dist, expected, atol=0.05)
-
-
-def test_multiple_molecules():
-    """Test two molecules, one crossing a boundary, one inside."""
-    mol1 = make_molecule(shift=(9.5, 0, 0))  # Wrapped
-    mol2 = make_molecule(shift=(5, 5, 5))  # Inside
-    atoms = mol1 + mol2
-    atoms.set_cell([10, 10, 10])
-    atoms.set_pbc([True, True, True])
-
-    atoms_unwrapped = unwrap_structures(atoms.copy())
-
-    # Check both H-H distances are ~1.1
-    d1 = atoms_unwrapped.get_distance(0, 1, mic=True)
-    d2 = atoms_unwrapped.get_distance(2, 3, mic=True)
-
-    assert np.isclose(d1, 1.1, atol=0.05)
-    assert np.isclose(d2, 1.1, atol=0.05)
-
-
-def test_wrapping_corner_case():
-    """Molecule across 3D corner (x, y, z all wrapped)."""
-    # Place one atom near box corner, the other offset by 1.1 Å along the diagonal
-    a1 = np.array([9.9, 9.9, 9.9])
-    bond_length = 1.1 / np.sqrt(3)
-    a2 = (a1 + np.array([bond_length] * 3)) % 10.0  # wrapped across corner
-
-    atoms = Atoms("H2", positions=[a1, a2], cell=[10, 10, 10], pbc=True)
-
-    atoms_unwrapped = unwrap_structures(atoms.copy())
-    dist = atoms_unwrapped.get_distance(0, 1, mic=True)
-
-    assert np.isclose(dist, 1.1, atol=0.05)
-
-
 def test_no_pbc():
     """Ensure no changes when system has no PBC."""
     atoms = make_molecule()
