### 2025-07-16 e7a5e10
test cyclic molecules
