### 2025-07-16 bb0ae4e
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
