+++ rdkit2ase/rdkit2x.py
@@ -35,7 +35,7 @@ def rdkit2ase(mol: Chem.Mol, seed: int = 42) -> ase.Atoms:
     """
     if mol.GetNumAtoms() == 0:
         return ase.Atoms()
-    
+
     smiles = Chem.MolToSmiles(mol)
     mol = Chem.AddHs(mol)
     charges = [atom.GetFormalCharge() for atom in mol.GetAtoms()]
