### 2025-07-16 2ccc8a7
v0.2.11 (#55)
* bump version

* empty molecule edge cases

* rename `unwrap_structures`

* update docstring

* add `scale` parameter

* add TODO

* simplify unwrap code

* remove tests that did nothing

* test cyclic molecules

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* address comments

* include node mapping for non-sequential data

* add test

* add one more test

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
