+++ tests/test_utils.py
@@ -1,5 +1,6 @@
 import sys
 
+import networkx as nx
 import numpy as np
 import pytest
 import rdkit.Chem
@@ -10,7 +11,7 @@ from rdkit2ase.utils import (
     find_connected_components,
     rdkit_determine_bonds,
     suggestions2networkx,
-    unwrap_molecule,
+    unwrap_structures,
 )
 
 
@@ -36,6 +37,91 @@ def ec_emc_li_pf6():
     )
 
 
+@pytest.mark.parametrize("scale", [1.0, 0.5, 2.0])
+def test_unwrap_structures(scale):
+    hexane = rdkit2ase.smiles2conformers("CCCCCC", numConfs=1)
+    box = rdkit2ase.pack(
+        data=[hexane],
+        counts=[10],
+        density=800,
+        packmol="packmol.jl",
+    )
+    shifted_box = box.copy()
+    shifted_box.set_positions(
+        shifted_box.get_positions() + scale * shifted_box.get_cell().diagonal()
+    )
+    shifted_box.wrap()  # Wrap back into PBC
+
+    # check that the movement broke some bonds
+    off = 0
+    graph = rdkit2ase.ase2networkx(shifted_box)
+    assert len(list(nx.connected_components(graph))) == 10
+    for i, j in graph.edges():
+        off += bool(shifted_box.get_distance(i, j, mic=False) > 1.6)
+    assert off > 0
+
+    # check the initial box has no broken bonds
+    off = 0
+    graph = rdkit2ase.ase2networkx(box)
+    assert len(list(nx.connected_components(graph))) == 10
+    for i, j in graph.edges():
+        off += bool(box.get_distance(i, j, mic=False) > 1.6)
+    assert off == 0
+
+    # Unwrap the structures and assert that bonds are restored
+    unwrapped_atoms = unwrap_structures(shifted_box)
+    graph = rdkit2ase.ase2networkx(unwrapped_atoms)
+    for i, j in graph.edges():
+        assert unwrapped_atoms.get_distance(i, j, mic=False) < 1.6
+
+    # assert edges stay the same / indices remain the same
+    graph_unwrapped = rdkit2ase.ase2networkx(unwrapped_atoms)
+    graph_box = rdkit2ase.ase2networkx(box)
+    assert nx.is_isomorphic(graph_unwrapped, graph_box)
+
+
+@pytest.mark.parametrize("scale", [1.0, 0.5, 2.0])
+def test_unwrap_ring_molecules(scale):
+    # Create a single benzene conformer
+    benzene = rdkit2ase.smiles2conformers("c1ccccc1", numConfs=1)
+
+    # Pack a box with 10 benzene molecules
+    box = rdkit2ase.pack(
+        data=[benzene],
+        counts=[10],
+        density=800,
+        packmol="packmol.jl",
+    )
+
+    # Copy and shift box to break molecules across PBC
+    shifted_box = box.copy()
+    shifted_box.set_positions(
+        shifted_box.get_positions() + scale * shifted_box.get_cell().diagonal()
+    )
+    shifted_box.wrap()  # Wraps back into the periodic box
+    # NOTE: the wrap seems to break bonds already!
+    # and yes, packmol places 2 hydrogens across the PBC boundary!
+
+    # Verify that some bonds are broken (distances > 1.6 Å without MIC)
+    graph = rdkit2ase.ase2networkx(shifted_box)
+    assert len(list(nx.connected_components(graph))) == 10  # Sanity check
+    broken_bonds = 0
[…]
