### 2025-07-21 d383c5d
add `get_center_of_mass`
