+++ tests/test_com.py
@@ -1,0 +1,112 @@
+import rdkit2ase
+import numpy.testing as npt
+import numpy as np
+import ase
+import pytest
+
+def test_com_isolated():
+    ethanol = rdkit2ase.smiles2atoms("CCO")
+    ethanol.positions += np.array([1.0, 0.0, 0.0])  # Shift to ensure non-zero COM
+    com = rdkit2ase.get_center_of_mass(ethanol)
+    npt.assert_allclose(com.get_center_of_mass(), ethanol.get_center_of_mass(), atol=1e-6)
+    npt.assert_allclose(com.get_positions()[0], ethanol.get_center_of_mass(), atol=1e-6)
+
+def test_com_box():
+    ethanol = rdkit2ase.smiles2atoms("CCO")
+    ethanol_com = ethanol.get_center_of_mass()
+    
+    box = ase.Atoms()
+    # add 5 ethanol molecules, shift them each so they don't overlap and then ensure the COM from rdkit2ase.get_center_of_mass(ethanol) matched the COM of each individual ethanol
+    shifts = np.array([[4 * i, 0, 0] for i in range(5)])
+    for shift in shifts:
+        ethanol_copy = ethanol.copy()
+        ethanol_copy.positions += shift
+        box += ethanol_copy
+    box_com = rdkit2ase.get_center_of_mass(box)
+    npt.assert_allclose(box_com.get_positions(), [ethanol_com + shift for shift in shifts], atol=1e-6)
+
+
+def test_com_explicitly_wrapped_molecule():
+    """
+    Tests the COM calculation by creating a box with two water molecules,
+    one of which is explicitly wrapped across a periodic boundary.
+    """
+    # 1. Setup a periodic box and a water molecule template
+    cell_dim = 7.0
+    box = ase.Atoms(cell=np.eye(3) * cell_dim, pbc=True)
+    water = rdkit2ase.smiles2atoms("O")
+    water.positions -= water.get_center_of_mass()
+    assert np.allclose(water.get_center_of_mass(), [0.0, 0.0, 0.0])
+    # 2. Define the target COM positions for two water molecules
+    # The second molecule is placed near the x-boundary to force wrapping.
+    expected_coms = np.array([
+        [cell_dim * 0.5, cell_dim * 0.5, cell_dim * 0.5],
+        [cell_dim * 0.9, cell_dim * 0.5, cell_dim * 0.5]
+    ])
+
+    # 3. Create the two molecules and add them to the box
+    water1 = water.copy()
+    water1.positions += expected_coms[0]
+    
+    water2 = water.copy()
+    water2.positions += expected_coms[1]
+
+    box += water1
+    box += water2
+
+    # 4. Wrap the box. This will break water2 across the boundary.
+    box.wrap()
+
+    # 5. Calculate COMs using the function under test
+    com_atoms = rdkit2ase.get_center_of_mass(box)
+    calculated_coms = com_atoms.get_positions()
+
+    # 6. Compare results. The order is not guaranteed, so we sort by a stable
+    # axis (x-coordinate in this case) before comparing.
+    assert len(calculated_coms) == len(expected_coms)
+    
+    sort_idx_calc = np.argsort(calculated_coms[:, 0])
+    sort_idx_exp = np.argsort(expected_coms[:, 0])
+    
+    sorted_calculated = calculated_coms[sort_idx_calc]
+    sorted_expected = expected_coms[sort_idx_exp]
+
+    npt.assert_allclose(sorted_calculated, sorted_expected, atol=1e-6)
+
+@pytest.mark.parametrize("shift", [0, 5])
+def test_com_with_packed_system(shift):
+    """
+    Tests the molecule identification part of the COM function on a dense,
[…]
