+++ tests/test_com.py
@@ -1,31 +1,40 @@
-import rdkit2ase
-import numpy.testing as npt
-import numpy as np
 import ase
+import numpy as np
+import numpy.testing as npt
 import pytest
+
+import rdkit2ase
+
 
 def test_com_isolated():
     ethanol = rdkit2ase.smiles2atoms("CCO")
     ethanol.positions += np.array([1.0, 0.0, 0.0])  # Shift to ensure non-zero COM
     com = rdkit2ase.get_center_of_mass(ethanol)
-    npt.assert_allclose(com.get_center_of_mass(), ethanol.get_center_of_mass(), atol=1e-6)
+    npt.assert_allclose(
+        com.get_center_of_mass(), ethanol.get_center_of_mass(), atol=1e-6
+    )
     npt.assert_allclose(com.get_positions()[0], ethanol.get_center_of_mass(), atol=1e-6)
     npt.assert_allclose(ethanol.cell, com.get_cell(), atol=1e-6)
     npt.assert_allclose(ethanol.pbc, com.pbc, atol=1e-6)
 
+
 def test_com_box():
     ethanol = rdkit2ase.smiles2atoms("CCO")
     ethanol_com = ethanol.get_center_of_mass()
-    
+
     box = ase.Atoms()
-    # add 5 ethanol molecules, shift them each so they don't overlap and then ensure the COM from rdkit2ase.get_center_of_mass(ethanol) matched the COM of each individual ethanol
+    # add 5 ethanol molecules, shift them each so they don't overlap and then
+    # ensure the COM from rdkit2ase.get_center_of_mass(ethanol)
+    # matched the COM of each individual ethanol
     shifts = np.array([[4 * i, 0, 0] for i in range(5)])
     for shift in shifts:
         ethanol_copy = ethanol.copy()
         ethanol_copy.positions += shift
         box += ethanol_copy
     box_com = rdkit2ase.get_center_of_mass(box)
-    npt.assert_allclose(box_com.get_positions(), [ethanol_com + shift for shift in shifts], atol=1e-6)
+    npt.assert_allclose(
+        box_com.get_positions(), [ethanol_com + shift for shift in shifts], atol=1e-6
+    )
     npt.assert_allclose(box.cell, box_com.get_cell(), atol=1e-6)
     npt.assert_allclose(box.pbc, box_com.pbc, atol=1e-6)
 
@@ -43,15 +52,17 @@ def test_com_explicitly_wrapped_molecule():
     assert np.allclose(water.get_center_of_mass(), [0.0, 0.0, 0.0])
     # 2. Define the target COM positions for two water molecules
     # The second molecule is placed near the x-boundary to force wrapping.
-    expected_coms = np.array([
-        [cell_dim * 0.5, cell_dim * 0.5, cell_dim * 0.5],
-        [cell_dim * 0.9, cell_dim * 0.5, cell_dim * 0.5]
-    ])
+    expected_coms = np.array(
+        [
+            [cell_dim * 0.5, cell_dim * 0.5, cell_dim * 0.5],
+            [cell_dim * 0.9, cell_dim * 0.5, cell_dim * 0.5],
+        ]
+    )
 
     # 3. Create the two molecules and add them to the box
     water1 = water.copy()
     water1.positions += expected_coms[0]
-    
+
     water2 = water.copy()
     water2.positions += expected_coms[1]
 
@@ -68,14 +79,15 @@ def test_com_explicitly_wrapped_molecule():
     # 6. Compare results. The order is not guaranteed, so we sort by a stable
     # axis (x-coordinate in this case) before comparing.
     assert len(calculated_coms) == len(expected_coms)
-    
+
     sort_idx_calc = np.argsort(calculated_coms[:, 0])
[…]
