+++ rdkit2ase/__init__.py
@@ -1,4 +1,5 @@
 from rdkit2ase.ase2x import ase2networkx, ase2rdkit
+from rdkit2ase.com import get_center_of_mass
 from rdkit2ase.compress import compress
 from rdkit2ase.networkx2x import networkx2ase, networkx2rdkit
 from rdkit2ase.pack import pack
@@ -6,7 +7,6 @@ from rdkit2ase.rdkit2x import rdkit2ase, rdkit2networkx
 from rdkit2ase.smiles2x import smiles2atoms, smiles2conformers
 from rdkit2ase.substructure import get_substructures, iter_fragments, match_substructure
 from rdkit2ase.utils import unwrap_structures
-from rdkit2ase.com import get_center_of_mass
 
 __all__ = [
     "ase2rdkit",
