### 2025-07-21 524b390
get_center_of_mass (#56)
* add `get_center_of_mass`

* bump version

* test for pbc / cell

* lint

* docs

* use species per type
