### 2025-07-21 0926ea7
convert numpy generics to native python types
