### 2025-07-21 b15bc43
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
