+++ rdkit2ase/ase2x.py
@@ -167,9 +167,7 @@ def ase2networkx(
             (int(i), int(j), float(bond_order) if bond_order is not None else None)
             for i, j, bond_order in connectivity
         ]
-        return _create_graph_from_connectivity(
-            atoms, connectivity, charges
-        )
+        return _create_graph_from_connectivity(atoms, connectivity, charges)
 
     connectivity_matrix, non_bonding_atomic_numbers = _compute_connectivity_matrix(
         atoms, scale, pbc
