### 2025-07-21 85c3ba0
Generics in connectivity (#58)
* convert numpy generics to native python types

* add addtional kwargs to unwrap_strucutre fixes accept kwargs for `ase2networkx` in `unwrap_structures`
Fixes #57

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* tests

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
