### 2025-07-25 #8530 @PythonFZ
@lucydot First of all have a good time off the internet.

I've looked through the reviewers and here might be some that could have something to say on the manuscript:
- https://reviewers.joss.theoj.org/reviewers/14401
- https://reviewers.joss.theoj.org/reviewers/3490
- https://reviewers.joss.theoj.org/reviewers/18986
- https://reviewers.joss.theoj.org/reviewers/18851
- https://reviewers.joss.theoj.org/reviewers/3334
- https://reviewers.joss.theoj.org/reviewers/1349
