### 2025-07-28 d52e59d
Add advanced atom selection and visualization functions
- Add select_atoms_flat_unique() with mapped SMILES support and sophisticated hydrogen handling
- Add visualize_selected_molecules() for highlighting atom selections
- Include comprehensive tests covering all functionality
- Add documentation notebook with examples

🤖 Generated with [Claude Code](https://claude.ai/code)

Co-Authored-By: Claude <noreply@anthropic.com>
