+++ rdkit2ase/__init__.py
@@ -5,7 +5,13 @@ from rdkit2ase.networkx2x import networkx2ase, networkx2rdkit
 from rdkit2ase.pack import pack
 from rdkit2ase.rdkit2x import rdkit2ase, rdkit2networkx
 from rdkit2ase.smiles2x import smiles2atoms, smiles2conformers
-from rdkit2ase.substructure import get_substructures, iter_fragments, match_substructure
+from rdkit2ase.substructure import (
+    get_substructures,
+    iter_fragments,
+    match_substructure,
+    select_atoms_flat_unique,
+    visualize_selected_molecules,
+)
 from rdkit2ase.utils import unwrap_structures
 
 __all__ = [
@@ -17,6 +23,8 @@ __all__ = [
     "match_substructure",
     "get_substructures",
     "iter_fragments",
+    "select_atoms_flat_unique",
+    "visualize_selected_molecules",
     "unwrap_structures",
     #
     "ase2networkx",
