+++ rdkit2ase/substructure.py
@@ -1,6 +1,9 @@
+import typing as tp
+
 import ase
 from ase.build import separate
 from rdkit import Chem
+from rdkit.Chem import Draw
 
 from rdkit2ase.ase2x import ase2rdkit
 from rdkit2ase.utils import find_connected_components
@@ -119,3 +122,193 @@ def iter_fragments(atoms: ase.Atoms) -> list[ase.Atoms]:
     else:
         for molecule in separate(atoms):
             yield molecule
+
+
+def select_atoms_flat_unique(
+    mol: Chem.Mol,
+    smarts_or_smiles: str,
+    hydrogens: tp.Literal["include", "exclude", "isolated"] = "exclude",
+) -> list[int]:
+    """
+    Selects a unique list of atom indices in a molecule using SMARTS or mapped SMILES.
+    If the pattern contains atom maps (e.g., [C:1]), only the mapped atoms are returned.
+    Otherwise, all atoms in the matched substructure are returned.
+
+    Parameters
+    ----------
+    mol : Chem.Mol
+        RDKit molecule, which can contain explicit hydrogens.
+    smarts_or_smiles : str
+        SMARTS (e.g., "[F]") or SMILES with atom maps (e.g., "C1[C:1]OC(=[O:1])O1").
+    hydrogens : {"include", "exclude", "isolated"}, default "exclude"
+        How to handle hydrogens in the final returned list.
+        - "include": Include hydrogens attached to matched heavy atoms
+        - "exclude": Exclude all hydrogens from results (default)
+        - "isolated": Return only hydrogens attached to matched heavy atoms
+
+    Returns
+    -------
+    list[int]
+        A single, flat list of unique integer atom indices matching the criteria.
+
+    Raises
+    ------
+    ValueError
+        If the SMARTS/SMILES pattern is invalid.
+    """
+    patt = Chem.MolFromSmarts(smarts_or_smiles)
+
+    if not patt:
+        raise ValueError(f"Invalid SMARTS/SMILES: {smarts_or_smiles}")
+
+    # Check if the pattern has any mapped atoms.
+    mapped_pattern_indices = [
+        atom.GetIdx() for atom in patt.GetAtoms() if atom.GetAtomMapNum() > 0
+    ]
+
+    matches = mol.GetSubstructMatches(patt)
+
+    # 1. Get the core set of atoms. If the pattern is mapped, only use the
+    #    indices corresponding to the mapped atoms. Otherwise, use all atoms.
+    core_atom_indices = set()
+    if mapped_pattern_indices:
+        # Only collect indices from the molecule that correspond to a mapped atom in the pattern.
+        for match_tuple in matches:
+            for pattern_idx in mapped_pattern_indices:
+                core_atom_indices.add(match_tuple[pattern_idx])
+    else:
+        # Original behavior: get all atoms from all matches if no maps are present.
+        core_atom_indices = set(idx for match_tuple in matches for idx in match_tuple)
+
+    # 2. Handle the `hydrogens` parameter based on this core set of atoms.
+    if hydrogens == "include":
+        final_indices = set(core_atom_indices)
+        for idx in core_atom_indices:
+            atom = mol.GetAtomWithIdx(idx)
+            if atom.GetAtomicNum() != 1:
+                for neighbor in atom.GetNeighbors():
+                    if neighbor.GetAtomicNum() == 1:
+                        final_indices.add(neighbor.GetIdx())
+        return sorted(list(final_indices))
+
+    elif hydrogens == "exclude":
+        heavy_only = {
+            idx
+            for idx in core_atom_indices
+            if mol.GetAtomWithIdx(idx).GetAtomicNum() != 1
+        }
[…]
