### 2025-08-04 3773cdb
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.12.0 → v0.12.7](https://github.com/astral-sh/ruff-pre-commit/compare/v0.12.0...v0.12.7)
