+++ .pre-commit-config.yaml
@@ -26,7 +26,7 @@ repos:
         additional_dependencies: ["tomli"]
   - repo: https://github.com/astral-sh/ruff-pre-commit
     # Ruff version.
-    rev: v0.12.0
+    rev: v0.12.7
     hooks:
       # Run the linter.
       - id: ruff
