### 2025-08-05 #8530 @lucydot
@colinbousige @vigneshS29 @csadorf - hello 👋  - are you able to review this submission to JOSS?
