### 2025-08-09 fbb68bc
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
