### 2025-08-09 90cb7bf
include `select_atoms_grouped`
