### 2025-08-09 3d8b178
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
