### 2025-08-09 d616b28
update docs and tests
