### 2025-08-09 c60c8b7
Update rdkit2ase/substructure.py
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>
