### 2025-08-09 144a257
Update rdkit2ase/substructure.py
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>
