### 2025-08-09 ba600e4
Add advanced atom selection and visualization functions (#63)
* Add advanced atom selection and visualization functions

- Add select_atoms_flat_unique() with mapped SMILES support and sophisticated hydrogen handling
- Add visualize_selected_molecules() for highlighting atom selections
- Include comprehensive tests covering all functionality
- Add documentation notebook with examples

🤖 Generated with [Claude Code](https://claude.ai/code)

Co-Authored-By: Claude <noreply@anthropic.com>

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* i
[…]
