+++ tests/test_substructure.py
@@ -160,3 +160,201 @@ def test_bmim_bf4_no_info():
         assert sorted(box[match].get_chemical_symbols()) == sorted(
             bmim[0].get_chemical_symbols()
         )
+
+
+@pytest.fixture
+def ethanol_mol():
+    """Returns an RDKit molecule for ethanol (CCO) with explicit hydrogens."""
+    mol = Chem.MolFromSmiles("CCO")
+    return Chem.AddHs(mol)
+
+
+@pytest.fixture
+def toluene_mol():
+    """Returns an RDKit molecule for toluene (Cc1ccccc1) with explicit hydrogens."""
+    mol = Chem.MolFromSmiles("Cc1ccccc1")
+    return Chem.AddHs(mol)
+
+
+# =============================================================================
+# Test Cases for select_atoms_flat_unique
+# =============================================================================
+
+
+def test_select_carbons(ethanol_mol):
+    """Test selecting all carbon atoms."""
+    # Ethanol (CCO with Hs): C(0), C(1), O(2), H(3-8)
+    indices = rdkit2ase.select_atoms_flat_unique(ethanol_mol, "[#6]")
+    assert sorted(indices) == [0, 1]
+
+
+def test_select_oxygen(ethanol_mol):
+    """Test selecting the oxygen atom."""
+    indices = rdkit2ase.select_atoms_flat_unique(ethanol_mol, "[#8]")
+    assert sorted(indices) == [2]
+
+
+def test_no_matches(ethanol_mol):
+    """Test a SMARTS pattern that has no matches."""
+    indices = rdkit2ase.select_atoms_flat_unique(ethanol_mol, "[F]")  # Fluorine
+    assert indices == []
+
+
+# --- Hydrogen Handling Tests ---
+
+
+def test_hydrogens_excluded_by_default(ethanol_mol):
+    """Test that hydrogens are excluded by default."""
+    # C-O bond involves atoms 1 and 2. Hydrogens attached are not included.
+    indices = rdkit2ase.select_atoms_flat_unique(ethanol_mol, "CO")
+    assert sorted(indices) == [1, 2]
+
+
+def test_hydrogens_included(ethanol_mol):
+    """Test the 'include' option for hydrogens."""
+    # C-O bond (atoms 1, 2) plus attached hydrogens (atoms 6, 7, 8)
+    # H on O is atom 8. Hs on C(1) are 6, 7.
+    indices = rdkit2ase.select_atoms_flat_unique(ethanol_mol, "CO", hydrogens="include")
+    # Expected: C(1), O(2), H(6), H(7), H(8)
+    assert sorted(indices) == [1, 2, 6, 7, 8]
+
+
+def test_hydrogens_isolated(ethanol_mol):
+    """Test the 'isolated' option for hydrogens."""
+    # Select ONLY the hydrogens from the C-O match
+    indices = rdkit2ase.select_atoms_flat_unique(
+        ethanol_mol, "CO", hydrogens="isolated"
+    )
+    # Expected: H(6), H(7), H(8)
+    assert sorted(indices) == [6, 7, 8]
+
+
+def test_smarts_with_explicit_hydrogens(ethanol_mol):
+    """Test a SMARTS pattern that explicitly includes hydrogens."""
+    # Find all hydrogens attached to an oxygen
+    indices = rdkit2ase.select_atoms_flat_unique(
+        ethanol_mol, "[#8]-[H]", hydrogens="include"
+    )
+    # Expected: O(2), H(8)
+    assert sorted(indices) == [2, 8]
+
+    # Now isolate only the hydrogen from that match
+    h_indices = rdkit2ase.select_atoms_flat_unique(
[…]
