### 2025-08-11 f7157e3
fix order in substructure mapping
