+++ rdkit2ase/substructure.py
@@ -1,6 +1,7 @@
 import typing as tp
 
 import ase
+import matplotlib.pyplot as plt
 from ase.build import separate
 from rdkit import Chem
 from rdkit.Chem import Draw
@@ -135,8 +136,9 @@ def select_atoms_grouped(  # noqa: C901
     lists. Each inner list corresponds to a unique, disconnected molecular fragment
     that contained at least one match.
 
-    If the pattern contains atom maps (e.g., "[C:1]"), only the mapped atoms
-    are returned. Otherwise, all atoms in the matched substructures are returned.
+    If the pattern contains atom maps (e.g., "[C:1]", "[C:2]"), only the mapped atoms
+    are returned, ordered by their map numbers. Map numbers must be unique within
+    the pattern. Otherwise, all atoms in the matched substructures are returned.
 
     Parameters
     ----------
@@ -145,24 +147,26 @@ def select_atoms_grouped(  # noqa: C901
         explicit hydrogens.
     smarts_or_smiles : str
         SMARTS pattern (e.g., "[F]") or SMILES with atom maps
-        (e.g., "C1[C:1]OC(=[O:1])O1").
+        (e.g., "CC(=O)N[C:1]([C:2])[C:3](=O)[N:4]C"). When using mapped atoms,
+        map numbers must be unique.
     hydrogens : {'include', 'exclude', 'isolated'}, default='exclude'
         How to handle hydrogens in the final returned list for each group:
-        - 'include': Add hydrogens bonded to selected heavy atoms.
+        - 'include': Add hydrogens bonded to selected heavy atoms after each mapped atom.
         - 'exclude': Remove all hydrogens from the selection.
         - 'isolated': Return only the hydrogens that are bonded to selected heavy atoms.
 
     Returns
     -------
     list[list[int]]
-        A list of sorted integer lists. Each inner list contains the unique atom indices
-        for a matched, disconnected fragment. Fragments with no matches
-        are omitted from the output.
+        A list of integer lists. Each inner list contains the atom indices
+        for a matched, disconnected fragment. For mapped patterns, atoms are ordered
+        by their map numbers. Fragments with no matches are omitted from the output.
 
     Raises
     ------
     ValueError
-        If the provided SMARTS/SMILES pattern is invalid.
+        If the provided SMARTS/SMILES pattern is invalid or if atom map labels
+        are used multiple times within the same pattern.
 
     Examples
     --------
@@ -345,25 +349,51 @@ def select_atoms_flat_unique(
     return sorted(unique_indices)
 
 
-def visualize_selected_molecules(mol: Chem.Mol, a: list[int], b: list[int]):  # noqa: C901
+def visualize_selected_molecules(
+    mol: Chem.Mol, 
+    *args,
+    molsPerRow: int = 4,
+    subImgSize: tuple[int, int] = (200, 200),
+    legends: list[str] | None = None,
+    alpha: float = 0.5
+):  # noqa: C901
     """
-    Visualizes molecules that contain selected atoms, highlighting the selections.
+    Visualizes molecules with optional atom highlighting.
+    If no atom selections are provided, displays the molecule without highlights.
     Duplicate molecular structures will only be plotted once.
 
     Parameters
     ----------
     mol : Chem.Mol
         The RDKit molecule object, which may contain multiple fragments.
-    a : list[int]
-        A list of atom indices to be highlighted in the first color (e.g., pink).
-    b : list[int]
-        A list of atom indices to be highlighted in the second color (e.g., light blue).
+    *args : list[int]
+        Variable number of lists containing atom indices to be highlighted.
+        Each list will be assigned a different color from matplotlib's tab10 colormap.
+        If no arguments provided, displays the molecule without highlights.
[…]
