### 2025-08-11 f9db05a
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
