### 2025-08-11 9e781fe
fix vesin can't do different pbc on different axis Fixes #62
