### 2025-08-11 bc5a2e5
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
