### 2025-08-11 ba9c6d0
ensure select atoms grouped is in order (#65)
* test order

* fix order processing

* fix order in substructure mapping

* fix order, update plotting

* ignore ds_store file

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* lint

* update text

* fix vesin can't do different pbc on different axis
Fixes #62

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
