+++ rdkit2ase/substructure.py
@@ -1,6 +1,7 @@
 import typing as tp
 
 import ase
+import matplotlib.pyplot as plt
 from ase.build import separate
 from rdkit import Chem
 from rdkit.Chem import Draw
@@ -135,8 +136,9 @@ def select_atoms_grouped(  # noqa: C901
     lists. Each inner list corresponds to a unique, disconnected molecular fragment
     that contained at least one match.
 
-    If the pattern contains atom maps (e.g., "[C:1]"), only the mapped atoms
-    are returned. Otherwise, all atoms in the matched substructures are returned.
+    If the pattern contains atom maps (e.g., "[C:1]", "[C:2]"), only the mapped atoms
+    are returned, ordered by their map numbers. Map numbers must be unique within
+    the pattern. Otherwise, all atoms in the matched substructures are returned.
 
     Parameters
     ----------
@@ -145,24 +147,27 @@ def select_atoms_grouped(  # noqa: C901
         explicit hydrogens.
     smarts_or_smiles : str
         SMARTS pattern (e.g., "[F]") or SMILES with atom maps
-        (e.g., "C1[C:1]OC(=[O:1])O1").
+        (e.g., "CC(=O)N[C:1]([C:2])[C:3](=O)[N:4]C"). When using mapped atoms,
+        map numbers must be unique.
     hydrogens : {'include', 'exclude', 'isolated'}, default='exclude'
         How to handle hydrogens in the final returned list for each group:
-        - 'include': Add hydrogens bonded to selected heavy atoms.
+        - 'include': Add hydrogens bonded to selected heavy atoms after each
+          mapped atom.
         - 'exclude': Remove all hydrogens from the selection.
         - 'isolated': Return only the hydrogens that are bonded to selected heavy atoms.
 
     Returns
     -------
     list[list[int]]
-        A list of sorted integer lists. Each inner list contains the unique atom indices
-        for a matched, disconnected fragment. Fragments with no matches
-        are omitted from the output.
+        A list of integer lists. Each inner list contains the atom indices
+        for a matched, disconnected fragment. For mapped patterns, atoms are ordered
+        by their map numbers. Fragments with no matches are omitted from the output.
 
     Raises
     ------
     ValueError
-        If the provided SMARTS/SMILES pattern is invalid.
+        If the provided SMARTS/SMILES pattern is invalid or if atom map labels
+        are used multiple times within the same pattern.
 
     Examples
     --------
@@ -183,10 +188,28 @@ def select_atoms_grouped(  # noqa: C901
         patt = Chem.MolFromSmiles(smarts_or_smiles)
     if patt is None:
         raise ValueError(f"Invalid SMARTS/SMILES: {smarts_or_smiles}")
-    # Get mapped indices from the pattern, if any
-    mapped_pattern_indices = [
-        atom.GetIdx() for atom in patt.GetAtoms() if atom.GetAtomMapNum() > 0
-    ]
+
+    # Get mapped indices from the pattern, if any, and validate uniqueness
+    mapped_pattern_indices = []
+    atom_map_numbers = []
+    for atom in patt.GetAtoms():
+        if atom.GetAtomMapNum() > 0:
+            map_num = atom.GetAtomMapNum()
+            if map_num in atom_map_numbers:
+                raise ValueError(f"Label '{map_num}' is used multiple times")
+            atom_map_numbers.append(map_num)
+            mapped_pattern_indices.append(atom.GetIdx())
+
+    # If we have mapped atoms, we need to sort them by their
+    #  map numbers to preserve order
+    if mapped_pattern_indices:
+        # Create pairs of (map_number, pattern_index) and sort by map_number
+        map_index_pairs = [
[…]
