+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "rdkit2ase"
-version = "0.1.12"
+version = "0.1.13"
 description = "Interface between rdkit and ASE"
 readme = "README.md"
 license = "Apache-2.0"
@@ -10,6 +10,7 @@ authors = [
 ]
 dependencies = [
     "ase>=3.25.0",
+    "matplotlib>=3.10.1",
     "networkx>=3.4.2",
     "rdkit>=2024",
 ]
