### 2025-08-18 #8829 @editorialbot
molify: Molecular Structure Interface
**Submitting author:** @PythonFZ (<a href="http://orcid.org/0000-0002-6936-4692">Fabian Zills</a>)
**Repository:** https://github.com/zincware/rdkit2ase
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.2.2
