### 2025-09-05 zincware/rdkit2ase#69 @colinbousige
packmol error
Hello

I have started to review the article but can't run your code.

I have installed `packmol` with the recommended way, ie with `juliaup`:
```bash
curl -fsSL https://install.julialang.org | sh
```
then installed the `packmol` package
```julia
julia> import Pkg; Pkg.add("Packmol")
```

Then installed `rdkit2ase` with `pip install rdkit2ase`

Then, running this in `python` I get this error:

```python
from rdkit2ase import pack, smiles2conformers
water = smiles2conformers("O", numConfs=2)
print(water[0].info['connectivity'])

>>> [(0, 1, 1.0), (0, 2, 1.0)]

ethanol = smiles2conformers("CCO", numConfs=5)
density = 1000 # kg/m^3
box = pack([water, ethanol], [7, 5], density, packmol="packmol.jl")

>>> Traceback (most recent call last):
>>> File "<python-input-1>", line 6, in <module>
>>>     box = pack([water, ethanol], [7, 5], density, packmol="packmol.jl")
>>>   File "/opt/homebrew/lib/python3.13/site-packages/rdkit2ase/pack.py", line 214, in pack
>>>     packed_atoms: ase.Atoms = ase.io.read(tmpdir / f"mixture.{output_format}")
>>>                               ~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
>>>   File "/opt/homebrew/lib/python3.13/site-packages/ase/io/formats.py", line 787, in read
>>>     format = format or filetype(filename, read=isinstance(filename, str))
>>>                        ~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
>>>   File "/opt/homebrew/lib/python3.13/site-packages/ase/io/formats.py", line 962, in filetype
>>>     fd = open_with_compression(filename, 'rb')
>>>   File "/opt/homebrew/lib/python3.13/site-packages/ase/io/formats.py", line 591, in open_with_compression
>>>     return open(filename, mode)
>>> FileNotFoundError: [Errno 2] No such file or directory: '/var/folders/1x/5p2nhhqn0jl6ngzj131btpd40000gn/T/tmp7r99h1tc/mixture.pdb'
```

Tried this on python 3.13.3 and 3.12, same error.

C
