### 2025-09-05 #8829 @colinbousige
started reviewing, see non reproductible code in https://github.com/zincware/rdkit2ase/issues/69#issue-3386936828
