### 2025-09-08 ff69785
update readme to properly reflect the rdkit interface
