### 2025-09-08 zincware/rdkit2ase#69 @colinbousige
```bash
pip show rdkit2ase ase

Name: rdkit2ase
Version: 0.1.13
Summary: Interface between rdkit and ASE
Home-page:
Author:
Author-email: Fabian Zills <fzills@icp.uni-stuttgart.de>
License-Expression: Apache-2.0
Location: /opt/homebrew/lib/python3.13/site-packages
Requires: ase, matplotlib, networkx, rdkit
Required-by:
---
Name: ase
Version: 3.26.0
Summary: Atomic Simulation Environment
Home-page: https://ase-lib.org/
Author:
Author-email:
License-Expression: LGPL-2.1-or-later
Location: /opt/homebrew/lib/python3.13/site-packages
Requires: matplotlib, numpy, scipy
Required-by: rdkit2ase
```
