### 2025-09-08 3676902
Add Packmol Julia version detection and improve FileNotFoundError handling
Co-authored-by: PythonFZ <46721498+PythonFZ@users.noreply.github.com>
