### 2025-09-09 29c04fd
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
