+++ README.md
@@ -7,12 +7,12 @@
 
 # rdkit2ase - Interface between the rdkit and ASE package.
 
-Installation via `pip install rdkit2ase`. 
-For more information please visit the
+Installation via `pip install rdkit2ase`. For more information please visit the
 [documentation](https://zincware.github.io/rdkit2ase/).
 
-A common use case is to create 3D structures from SMILES strings.
-This can be achieved using the `rdkit2ase.rdkit2ase` function.
+A common use case is to create 3D structures from SMILES strings. This can be
+achieved using the `rdkit2ase.rdkit2ase` function.
+
 ```py
 import ase
 from rdkit import Chem
@@ -22,6 +22,7 @@ mol = Chem.MolFromSmiles("O")
 atoms: ase.Atoms = rdkit2ase(mol)
 mol = ase2rdkit(atoms)
 ```
+
 Because this is such a common use case, there is a convenience function
 `rdkit2ase.smiles2atoms` that combines the two steps.
 
@@ -36,8 +37,9 @@ print(atoms)
 ```
 
 ## Packmol Interface
-Give the molecular units, you can build periodic boxes with a given density using
-the `rdkit2ase.pack` function.
+
+Give the molecular units, you can build periodic boxes with a given density
+using the `rdkit2ase.pack` function.
 
 If you have [packmol](https://github.com/m3g/packmol) (at least `v20.15.0`) you
 can use the rdkit2ase interface as follows:
@@ -53,4 +55,5 @@ print(box)
 >>> Atoms(symbols='C10H44O12', pbc=True, cell=[8.4, 8.4, 8.4])
 ```
 
-Many additional features are described in the [documentation](https://zincware.github.io/rdkit2ase/).
+Many additional features are described in the
+[documentation](https://zincware.github.io/rdkit2ase/).
