### 2025-09-09 3724a9d
update README to include all requried import statemtns (#70)
* update README to include all requried import statemtns

* update readme to properly reflect the rdkit interface

* update docs

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* nitpick comments

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
