+++ rdkit2ase/utils.py
@@ -1,5 +1,6 @@
 import io
 from collections import defaultdict
+from unittest import result
 
 import ase.io
 import ase.units
@@ -8,7 +9,7 @@ import numpy as np
 import rdkit.Chem
 import rdkit.Chem.rdDetermineBonds
 from rdkit import Chem
-
+import subprocess
 
 def bond_type_from_order(order):
     if order == 1.0:
@@ -220,3 +221,29 @@ def suggestions2networkx(smiles: list[str]) -> list[nx.Graph]:
         mol = Chem.AddHs(mol)
         mols.append(mol)
     return [rdkit2networkx(mol) for mol in mols]
+
+
+def get_packmol_julia_version() -> str:
+    """Get the Packmol version when using Julia.
+    
+    Raises
+    ------
+    RuntimeError
+        If the Packmol version cannot be retrieved.
+    """
+    try:
+        result = subprocess.run(
+            ["julia", "-e", "import Pkg; Pkg.status(\"Packmol\")"],
+            capture_output=True,
+            text=True,
+            check=True,
+        )
+        for line in result.stdout.splitlines():
+            if "Packmol" in line:
+                parts = line.split()
+                if len(parts) >= 3:
+                    return parts[2]
+        raise RuntimeError("Failed to get Packmol version via Julia")
+        
+    except subprocess.CalledProcessError as e:
+        raise RuntimeError("Failed to get Packmol version via Julia") from e
