### 2025-09-09 05bd6e7
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
