### 2025-09-09 87de613
better error handling
