### 2025-09-09 3257630
remove unused array keys
