### 2025-09-09 bba5751
allow building rectangular boxes
