### 2025-09-09 f6f9e2a
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
