+++ rdkit2ase/pack.py
@@ -188,8 +188,8 @@ def pack(
         WARNING: Do not use "xyz". This might cause issues and
         is only implemented for debugging purposes.
     ratio : tuple[float, float, float], optional
-        The aspect ratio of the simulation box,
-        by default (1.0, 1.0, 1.0) for a cubic box.
+        Box aspect ratio (a:b:c). Must be three positive, finite numbers.
+        Defaults to (1.0, 1.0, 1.0) for a cubic box.
 
     Returns
     -------
@@ -208,14 +208,16 @@ def pack(
     """
     selected_images = _select_conformers(data, counts, seed)
     cell = calculate_box_dimensions(images=selected_images, density=density)
-    packmol_input = _generate_packmol_input(
-        selected_images, cell, tolerance, seed, output_format, pbc
-    )
+
     # Adjust cell dimensions according to ratio while keeping volume unchanged
     original_volume = np.prod(cell)
     ratio_product = np.prod(ratio)
     scale = original_volume ** (1 / 3) / (ratio_product ** (1 / 3))
-    cell = [scale * r for r in ratio]
+    cell = [float(scale * r) for r in ratio]
+
+    packmol_input = _generate_packmol_input(
+        selected_images, cell, tolerance, seed, output_format, pbc
+    )
 
     with tempfile.TemporaryDirectory() as tmpdir_str:
         tmpdir = pathlib.Path(tmpdir_str)
