### 2025-10-01 #8829 @csadorf
While the paper provides a clear description of the purpose, the documentation (including the GitHub README) lacks such an explanation. The documentation could benefit from a more concise description of the scope and purpose of the package. Users visiting the repository should immediately understand what rdkit2ase does without needing to read the full paper.

The author clearly describes how this software relates to its dependencies and how it serves as a link between different software packages, but it lacks a description of any alternative solutions to the same problem. A brief comparison with existing tools (e.g., OpenBabel, RDKit's built-in conversion capabilities) would help users understand the unique value proposition of this package.

Given the narrow scope, I am wondering whether it would have been feasible or advantageous to simply contribute this function to rdkit directly. On the flip side, there appears to be some scope creep. Functions like [`get_centers_of_mass()`](https://zincware.github.io/rdkit2ase/modules.html#rdkit2ase.get_centers_of_mass) would be more appropriately implemented directly in ASE. Other examples of functions that appear out-of-scope are [`match_substructure()`](https://zincware.github.io/rdkit2ase/modules.html#rdkit2ase.match_substructure) and [`visualize_selected_molecules()`](https://zincware.github.io/rdkit2ase/modules.html#rdkit2ase.visualize_selected_molecules).

The paper's introduction provides clear motivation for why a user would want to be able to convert SMILES into ASE structures (or any of the other supported formats). However, the motivation is very specific to the stated application domain (Machine-Learned Interatomic Potentials). It seems like the actual function of the software would be much more widely applicable and it is not immediately clear why the author is restricting the motivation so narrowly. The broader applications in computational chemistry, materials science, and drug discovery should be acknowledged.

The package has straightforward installation instructions, but the documentation could be clearer about system dependencies. The code appears to be well maintained and developed overall. The package includes comprehensive testing that runs within CI on a comprehensive test matrix. The conversion functions appear well-designed with clear input/output specifications.

I was able to install the software following the instructions and run some of the examples without any problems.

I believe the paper is likely acceptable with revision. My main concern is that the package's scope is somewhat ill-defined. Some functions could be more appropriately contributed to either ASE or RDKit, while others fall entirely outside the expected scope. The package name suggests an extremely narrow focus on converting RDKit formats to ASE, but the documentation and functionality extend far beyond this core purpose and is in fact prominently showcasing other use cases, e.g., SMILES parsing.
