### 2025-10-16 #8829 @PythonFZ
Thank you @colinbousige and @csadorf for the review, and @lucydot for managing the review process.

Please excuse the delay in my response — I will be working on this next week.

I plan to address most of the comments individually, but I would like your feedback on one fundamental issue:

> The package name suggests an extremely narrow focus on converting RDKit formats to ASE.

Would it be acceptable to rename both the manuscript title and the package to better emphasize the broader focus on working with molecular structures in ASE, NetworkX, and RDKit?

This change would also address the concern that

> [...] it is not immediately clear why the author is restricting the motivation so narrowly,

by making it clearer that this is a more general-purpose tool for molecular data. It would also respond to the comment that

> [...] on the flip side, there appears to be some scope creep,

since what currently appears as scope creep is, in my opinion, a limitation of ASE - which, understandably, does not define molecular connectivity. Therefore, adding much of the features here to ASE would not be align with the packages design.

These features are already implemented in the existing rdkit2ase package, so the main changes would involve renaming the package and updating the focus throughout the documentation, README, and manuscript.

The new name I have in mind is `molbridge` (but suggestions are very welcome!).
