### 2025-10-23 484646f
refactor: clarify bond order determination architecture
Separates connectivity detection from bond order determination:
- ase2networkx() now only determines which atoms are bonded
- networkx2rdkit() handles all bond order determination
- ase2rdkit() simplified to a clean chain of the above

This resolves confusion around None bond orders and provides a
clearer API with proper separation of concerns.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
