### 2025-10-23 2300b8d
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
