+++ tests/integration/test_ase2x.py
@@ -116,10 +116,12 @@ def test_ase2networkx_guess_connectivity(atoms_and_connectivity):
     # To get bond orders, we need to pass through networkx2rdkit
     graph = rdkit2ase.ase2networkx(atoms)
     from rdkit2ase import networkx2rdkit
+
     # This will determine bond orders
     mol = networkx2rdkit(graph, suggestions=[])
     # Convert back to verify bond orders were determined
     from rdkit2ase import rdkit2networkx
+
     graph_with_orders = rdkit2networkx(mol)
     for i, j, bond_order in connectivity:
         assert graph_with_orders.has_edge(i, j)
@@ -134,10 +136,12 @@ def test_ase2networkx_smiles_connectivity(atoms_and_connectivity):
     # To get bond orders with SMILES suggestions, pass through networkx2rdkit
     graph = rdkit2ase.ase2networkx(atoms)
     from rdkit2ase import networkx2rdkit
+
     # This will determine bond orders using the SMILES suggestion
     mol = networkx2rdkit(graph, suggestions=[smiles])
     # Convert back to verify bond orders were determined
     from rdkit2ase import rdkit2networkx
+
     graph_with_orders = rdkit2networkx(mol)
     for i, j, bond_order in connectivity:
         assert graph_with_orders.has_edge(i, j)
@@ -234,16 +238,20 @@ def test_ase2networkx_ec_emc_li_pf6_guess_connectivity(ec_emc_li_pf6):
     # To get bond orders, we need to pass through networkx2rdkit
     graph = rdkit2ase.ase2networkx(ec_emc_li_pf6)
     from rdkit2ase import networkx2rdkit
+
     # This will determine bond orders
     mol = networkx2rdkit(graph, suggestions=[])
     # Convert back to verify bond orders were determined
     from rdkit2ase import rdkit2networkx
+
     graph_with_orders = rdkit2networkx(mol)
     for i, j, bond_order in connectivity:
         assert graph_with_orders.has_edge(i, j)
         assert graph_with_orders.edges[i, j]["bond_order"] == bond_order
 
-    li_nodes = [d for _, d in graph_with_orders.nodes(data=True) if d["atomic_number"] == 3]
+    li_nodes = [
+        d for _, d in graph_with_orders.nodes(data=True) if d["atomic_number"] == 3
+    ]
     assert len(li_nodes) == 3
     for d in li_nodes:
         assert d["charge"] == 1
