### 2025-10-23 8754f70
test: refactor test suite for clarity and coverage
Removes hacky roundtrip tests and adds comprehensive direct tests:

**Removed:**
- test_ase2networkx_guess_connectivity (hacky roundtrip)
- test_ase2networkx_smiles_connectivity (hacky roundtrip)
- test_ase2networkx_ec_emc_li_pf6_guess_connectivity (hacky roundtrip)
- test_ase2rdkit_guess_connectivity (duplicate)

**Added to test_networkx2x.py:**
- test_networkx2rdkit_with_none_bond_orders (26 variants)
- test_networkx2rdkit_with_smiles_suggestions (26 variants)
- test_networkx2rdkit_preserves_input_graph
- test_networkx2rdkit_mixed_bond_order
[…]
