### 2025-10-23 5339f83
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
