### 2025-10-23 e977f6b
fix: automatically handle None bond orders in ase2rdkit (#77)
* fix: automatically handle None bond orders in ase2rdkit

When connectivity information exists but contains None bond orders,
ase2rdkit would fail with:
  ValueError: Edge (0, 1) is missing 'bond_order' attribute

This can happen when:
- Loading structures from external files with incomplete connectivity
- Processing packed systems where connectivity wasn't properly set
- Using select_atoms_grouped on systems with corrupted metadata

Changes:
- Modified ase2rdkit() to detect None bond orders in connectivity
- Automatically removes 
[…]
