+++ tests/integration/test_ase2x.py
@@ -109,26 +109,6 @@ def test_ase2networkx_no_connectivity(atoms_and_connectivity):
 
 
 @pytest.mark.parametrize("atoms_and_connectivity", SMILES_LIST, indirect=True)
-def test_ase2networkx_guess_connectivity(atoms_and_connectivity):
-    _, atoms, connectivity = atoms_and_connectivity
-    atoms.info.pop("connectivity")
-    graph = rdkit2ase.ase2networkx(atoms, suggestions=[])
-    for i, j, bond_order in connectivity:
-        assert graph.has_edge(i, j)
-        assert graph.edges[i, j]["bond_order"] == bond_order
-
-
-@pytest.mark.parametrize("atoms_and_connectivity", SMILES_LIST, indirect=True)
-def test_ase2networkx_smiles_connectivity(atoms_and_connectivity):
-    smiles, atoms, connectivity = atoms_and_connectivity
-    atoms.info.pop("connectivity")
-    graph = rdkit2ase.ase2networkx(atoms, suggestions=[smiles])
-    for i, j, bond_order in connectivity:
-        assert graph.has_edge(i, j)
-        assert graph.edges[i, j]["bond_order"] == bond_order
-
-
-@pytest.mark.parametrize("atoms_and_connectivity", SMILES_LIST, indirect=True)
 def test_ase2rdkit(atoms_and_connectivity):
     smiles, atoms, connectivity = atoms_and_connectivity
     mol = rdkit2ase.ase2rdkit(atoms)
@@ -152,20 +132,7 @@ def test_ase2rdkit(atoms_and_connectivity):
 
 @pytest.mark.parametrize("atoms_and_connectivity", SMILES_LIST, indirect=True)
 def test_ase2rdkit_no_connectivity(atoms_and_connectivity):
-    smiles, atoms, connectivity = atoms_and_connectivity
-    atoms.info.pop("connectivity")
-    if len(connectivity) > 0:
-        with pytest.raises(ValueError):
-            rdkit2ase.ase2rdkit(atoms)
-    else:
-        mol = rdkit2ase.ase2rdkit(atoms)
-        assert Chem.MolToSmiles(mol, canonical=True) == Chem.MolToSmiles(
-            Chem.AddHs(Chem.MolFromSmiles(smiles)), canonical=True
-        )
-
-
-@pytest.mark.parametrize("atoms_and_connectivity", SMILES_LIST, indirect=True)
-def test_ase2rdkit_guess_connectivity(atoms_and_connectivity):
+    """Test ase2rdkit determines bond orders when connectivity is not provided."""
     smiles, atoms, connectivity = atoms_and_connectivity
     atoms.info.pop("connectivity")
     mol = rdkit2ase.ase2rdkit(atoms, suggestions=[])
@@ -211,20 +178,6 @@ def test_ase2networkx_ec_emc_li_pf6_no_connectivity(ec_emc_li_pf6):
     # assert len(li_nodes) == 3
     # for d in li_nodes:
     #     assert d["charge"] == 1
-
-
-def test_ase2networkx_ec_emc_li_pf6_guess_connectivity(ec_emc_li_pf6):
-    connectivity = ec_emc_li_pf6.info.pop("connectivity")
-    ec_emc_li_pf6.set_initial_charges()
-    graph = rdkit2ase.ase2networkx(ec_emc_li_pf6, suggestions=[])
-    for i, j, bond_order in connectivity:
-        assert graph.has_edge(i, j)
-        assert graph.edges[i, j]["bond_order"] == bond_order
-
-    li_nodes = [d for _, d in graph.nodes(data=True) if d["atomic_number"] == 3]
[…]
