### 2025-10-23 880de1b
rename `rdkit2ase` to `molify`
