+++ tests/integration/test_ase2x.py
@@ -4,7 +4,7 @@ import ase
 import pytest
 from rdkit import Chem
 
-import rdkit2ase
+import molify
 
 
 class SMILES:
@@ -16,12 +16,12 @@ class SMILES:
 
 @pytest.fixture
 def ec_emc_li_pf6():
-    atoms_pf6 = rdkit2ase.smiles2conformers(SMILES.PF6, numConfs=10)
-    atoms_li = rdkit2ase.smiles2conformers(SMILES.Li, numConfs=10)
-    atoms_ec = rdkit2ase.smiles2conformers(SMILES.EC, numConfs=10)
-    atoms_emc = rdkit2ase.smiles2conformers(SMILES.EMC, numConfs=10)
+    atoms_pf6 = molify.smiles2conformers(SMILES.PF6, numConfs=10)
+    atoms_li = molify.smiles2conformers(SMILES.Li, numConfs=10)
+    atoms_ec = molify.smiles2conformers(SMILES.EC, numConfs=10)
+    atoms_emc = molify.smiles2conformers(SMILES.EMC, numConfs=10)
 
-    return rdkit2ase.pack(
+    return molify.pack(
         data=[atoms_pf6, atoms_li, atoms_ec, atoms_emc],
         counts=[3, 3, 8, 12],
         density=1400,
@@ -84,7 +84,7 @@ SMILES_LIST = [
 @pytest.fixture
 def atoms_and_connectivity(request):
     smiles = request.param
-    atoms = rdkit2ase.smiles2atoms(smiles)
+    atoms = molify.smiles2atoms(smiles)
     connectivity = atoms.info["connectivity"]
     return smiles, atoms, connectivity
 
@@ -92,7 +92,7 @@ def atoms_and_connectivity(request):
 @pytest.mark.parametrize("atoms_and_connectivity", SMILES_LIST, indirect=True)
 def test_ase2networkx(atoms_and_connectivity):
     _, atoms, connectivity = atoms_and_connectivity
-    graph = rdkit2ase.ase2networkx(atoms)
+    graph = molify.ase2networkx(atoms)
     for i, j, bond_order in connectivity:
         assert graph.has_edge(i, j)
         assert graph.edges[i, j]["bond_order"] == bond_order
@@ -102,7 +102,7 @@ def test_ase2networkx(atoms_and_connectivity):
 def test_ase2networkx_no_connectivity(atoms_and_connectivity):
     _, atoms, connectivity = atoms_and_connectivity
     atoms.info.pop("connectivity")
-    graph = rdkit2ase.ase2networkx(atoms)
+    graph = molify.ase2networkx(atoms)
     for i, j, bond_order in connectivity:
         assert graph.has_edge(i, j)
         assert graph.edges[i, j]["bond_order"] is None
@@ -111,7 +111,7 @@ def test_ase2networkx_no_connectivity(atoms_and_connectivity):
 @pytest.mark.parametrize("atoms_and_connectivity", SMILES_LIST, indirect=True)
 def test_ase2rdkit(atoms_and_connectivity):
     smiles, atoms, connectivity = atoms_and_connectivity
-    mol = rdkit2ase.ase2rdkit(atoms)
+    mol = molify.ase2rdkit(atoms)
     assert Chem.MolToSmiles(mol, canonical=True) == Chem.MolToSmiles(
         Chem.AddHs(Chem.MolFromSmiles(smiles)), canonical=True
     )
@@ -135,7 +135,7 @@ def test_ase2rdkit_no_connectivity(atoms_and_connectivity):
     """Test ase2rdkit determines bond orders when connectivity is not provided."""
     smiles, atoms, connectivity = atoms_and_connectivity
     atoms.info.pop("connectivity")
-    mol = rdkit2ase.ase2rdkit(atoms, suggestions=[])
[…]
