+++ tests/conftest.py
@@ -1,14 +1,14 @@
 import ase
 import pytest
 
-import rdkit2ase
+import molify
 
 
 @pytest.fixture(scope="session")
 def ethanol_water():
-    ethanol = rdkit2ase.smiles2conformers("CCO", numConfs=100)
-    water = rdkit2ase.smiles2conformers("O", numConfs=100)
-    box = rdkit2ase.pack(
+    ethanol = molify.smiles2conformers("CCO", numConfs=100)
+    water = molify.smiles2conformers("O", numConfs=100)
+    box = molify.pack(
         [ethanol, water], counts=[2, 2], density=700, packmol="packmol.jl"
     )
     return box.copy()
@@ -16,14 +16,14 @@ def ethanol_water():
 
 @pytest.fixture(scope="session")
 def alanine_dipeptide() -> ase.Atoms:
-    """Alanine dipeptide conformer using rdkit2ase.smiles2conformers."""
-    return rdkit2ase.smiles2atoms("CC(=O)NC(C)C(=O)NC")
+    """Alanine dipeptide conformer using molify.smiles2conformers."""
+    return molify.smiles2atoms("CC(=O)NC(C)C(=O)NC")
 
 
 @pytest.fixture(scope="session")
 def alanine_dipeptide_box(alanine_dipeptide) -> ase.Atoms:
-    """Box of alanine dipeptide molecules using rdkit2ase.pack"""
-    box = rdkit2ase.pack(
+    """Box of alanine dipeptide molecules using molify.pack"""
+    box = molify.pack(
         [[alanine_dipeptide]], counts=[3], density=500, packmol="packmol.jl"
     )
     return box.copy()
