### 2025-10-23 19cb71b
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
