### 2025-10-23 43c2c96
add `__version__`
