### 2025-10-23 235ca40
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
