+++ tests/test_substructure.py
@@ -1,77 +1,77 @@
 import pytest
 from rdkit import Chem
 
-import rdkit2ase
+import molify
 
 
 def test_match_substructure():
-    atoms = rdkit2ase.smiles2atoms("CC(=O)O")
+    atoms = molify.smiles2atoms("CC(=O)O")
 
     # match CH3 fragment using smarts
-    match = rdkit2ase.match_substructure(atoms, smarts="[C]([H])([H])[H]")
+    match = molify.match_substructure(atoms, smarts="[C]([H])([H])[H]")
     assert match == ((0, 4, 5, 6),)
     assert atoms[match[0]].get_chemical_symbols() == ["C", "H", "H", "H"]
 
     # now match using a ase.Atoms object
 
-    ref = rdkit2ase.smiles2atoms("[C]([H])([H])[H]")
-    match = rdkit2ase.match_substructure(atoms, fragment=ref)
+    ref = molify.smiles2atoms("[C]([H])([H])[H]")
+    match = molify.match_substructure(atoms, fragment=ref)
     assert match == ((0, 4, 5, 6),)
 
     # now match using a Chem.Mol object
     ref_mol = Chem.MolFromSmarts("[C]([H])([H])[H]")
-    match = rdkit2ase.match_substructure(atoms, mol=ref_mol)
+    match = molify.match_substructure(atoms, mol=ref_mol)
     assert match == ((0, 4, 5, 6),)
 
     # check everything else raises TypeError
     with pytest.raises(TypeError):
-        rdkit2ase.match_substructure(atoms, 42)  # type: ignore[arg-type]
+        molify.match_substructure(atoms, 42)  # type: ignore[arg-type]
 
 
 @pytest.mark.parametrize("packmol", ["packmol.jl"])
 def test_match_substructur_box(packmol):
-    atoms = rdkit2ase.smiles2conformers("CC(=O)O", 1)
-    box = rdkit2ase.pack([atoms], counts=[3], packmol=packmol, density=0.5)
+    atoms = molify.smiles2conformers("CC(=O)O", 1)
+    box = molify.pack([atoms], counts=[3], packmol=packmol, density=0.5)
 
     # match CH3 fragment using smarts
-    match = rdkit2ase.match_substructure(box, smarts="[C]([H])([H])[H]")
+    match = molify.match_substructure(box, smarts="[C]([H])([H])[H]")
     assert match == ((0, 4, 5, 6), (8, 12, 13, 14), (16, 20, 21, 22))
     for m in match:
         assert box[m].get_chemical_symbols() == ["C", "H", "H", "H"]
 
     # now match using a ase.Atoms object
-    ref = rdkit2ase.smiles2atoms("[C]([H])([H])[H]")
-    match = rdkit2ase.match_substructure(box, fragment=ref)
+    ref = molify.smiles2atoms("[C]([H])([H])[H]")
+    match = molify.match_substructure(box, fragment=ref)
     assert match == ((0, 4, 5, 6), (8, 12, 13, 14), (16, 20, 21, 22))
 
     # now match using a Chem.Mol object
     ref_mol = Chem.MolFromSmarts("[C]([H])([H])[H]")
-    match = rdkit2ase.match_substructure(box, mol=ref_mol)
+    match = molify.match_substructure(box, mol=ref_mol)
     assert match == ((0, 4, 5, 6), (8, 12, 13, 14), (16, 20, 21, 22))
 
     # check everything else raises TypeError
     with pytest.raises(TypeError):
-        rdkit2ase.match_substructure(box, 42)
+        molify.match_substructure(box, 42)
 
 
 def test_get_substructure():
-    atoms = rdkit2ase.smiles2atoms("C(C(CO[N+](=O)[O-])O[N+](=O)[O-])O[N+](=O)[O-]")
[…]
