+++ tests/integration/test_networkx2x.py
@@ -2,7 +2,7 @@ import numpy.testing as npt
 import pytest
 from rdkit import Chem
 
-import rdkit2ase
+import molify
 
 
 class SMILES:
@@ -14,12 +14,12 @@ class SMILES:
 
 @pytest.fixture
 def ec_emc_li_pf6():
-    atoms_pf6 = rdkit2ase.smiles2conformers(SMILES.PF6, numConfs=10)
-    atoms_li = rdkit2ase.smiles2conformers(SMILES.Li, numConfs=10)
-    atoms_ec = rdkit2ase.smiles2conformers(SMILES.EC, numConfs=10)
-    atoms_emc = rdkit2ase.smiles2conformers(SMILES.EMC, numConfs=10)
+    atoms_pf6 = molify.smiles2conformers(SMILES.PF6, numConfs=10)
+    atoms_li = molify.smiles2conformers(SMILES.Li, numConfs=10)
+    atoms_ec = molify.smiles2conformers(SMILES.EC, numConfs=10)
+    atoms_emc = molify.smiles2conformers(SMILES.EMC, numConfs=10)
 
-    return rdkit2ase.pack(
+    return molify.pack(
         data=[atoms_pf6, atoms_li, atoms_ec, atoms_emc],
         counts=[3, 3, 8, 12],
         density=1400,
@@ -87,15 +87,15 @@ def normalize_connectivity(connectivity):
 @pytest.fixture
 def graph_smiles_atoms(request):
     smiles = request.param
-    atoms = rdkit2ase.smiles2atoms(smiles)
-    return rdkit2ase.ase2networkx(atoms), smiles, atoms
+    atoms = molify.smiles2atoms(smiles)
+    return molify.ase2networkx(atoms), smiles, atoms
 
 
 @pytest.mark.parametrize("graph_smiles_atoms", SMILES_LIST, indirect=True)
 def test_networkx2ase(graph_smiles_atoms):
     graph, smiles, atoms = graph_smiles_atoms
 
-    new_atoms = rdkit2ase.networkx2ase(graph)
+    new_atoms = molify.networkx2ase(graph)
     assert new_atoms == atoms
 
     new_connectivity = normalize_connectivity(new_atoms.info["connectivity"])
@@ -109,9 +109,9 @@ def test_networkx2ase(graph_smiles_atoms):
 
 def test_networkx2ase_ec_emc_li_pf6(ec_emc_li_pf6):
     atoms = ec_emc_li_pf6
-    graph = rdkit2ase.ase2networkx(atoms)
+    graph = molify.ase2networkx(atoms)
 
-    new_atoms = rdkit2ase.networkx2ase(graph)
+    new_atoms = molify.networkx2ase(graph)
 
     new_connectivity = normalize_connectivity(new_atoms.info["connectivity"])
     old_connectivity = normalize_connectivity(atoms.info["connectivity"])
@@ -130,7 +130,7 @@ def test_networkx2ase_ec_emc_li_pf6(ec_emc_li_pf6):
 def test_networkx2rdkit(graph_smiles_atoms):
     graph, smiles, atoms = graph_smiles_atoms
 
-    mol = rdkit2ase.networkx2rdkit(graph)
+    mol = molify.networkx2rdkit(graph)
     assert Chem.MolToSmiles(mol, canonical=True) == Chem.MolToSmiles(
         Chem.AddHs(Chem.MolFromSmiles(smiles)), canonical=True
     )
@@ -138,8 +138,8 @@ def test_networkx2rdkit(graph_smiles_atoms):
 
 def test_networkx2rdkit_ec_emc_li_pf6(ec_emc_li_pf6):
     atoms = ec_emc_li_pf6
[…]
