+++ docs/source/index.rst
@@ -1,26 +1,145 @@
 molify Documentation
 ====================
 
-molify provides a bridge between RDKit_ and ASE_.
-It further integrates features from Packmol_ and NetworkX_.
+**molify** - molecular structure conversion between RDKit, ASE, and NetworkX
+
+The molify package provides tools to convert molecular structures between
+RDKit_, ASE_ and NetworkX_.
+
+- RDKit is a versatile and popular cheminformatics library.
+
+- ASE allows interfacing with a wide range of atomistic simulation codes and reads/writes many file formats.
+
+- NetworkX is designed for the creation, manipulation, and study of complex networks.
+
+Key Features
+------------
+
+- **Bidirectional conversions** between RDKit Mol, ASE Atoms, and NetworkX Graph objects
+- **3D conformer generation** from SMILES strings using RDKit's embedding algorithms
+- **Molecular system building** with Packmol integration for creating solvated systems and mixtures
+- **Connectivity preservation** with automatic bond detection when needed
+
+Conversion Overview
+-------------------
+
+These bridging functions are illustrated below.
+
+.. mermaid::
+
+   graph LR
+      A[rdkit.Mol] -->|molify.rdkit2networkx| B[networkx.Graph]
+      B -->|molify.networkx2ase| C[ase.Atoms]
+      A -->|molify.rdkit2ase| C
+
+.. mermaid::
+
+   graph LR
+      A[ase.Atoms] -->|molify.ase2networkx| B[networkx.Graph]
+      B -->|molify.networkx2rdkit| C[rdkit.Mol]
+      A -->|molify.ase2rdkit| C
+
+Conversion Details
+------------------
+The rdkit package is built around the concept of molecules, with defined connections, bond orders and charges. Positional information is optional.
+In contrast, ASE Atoms objects are built around atomic positions and atomic numbers, with no inherent concept of bonds or connectivity.
+NetworkX Graphs are general-purpose graph structures with no predefined chemistry concepts.
+
+To store the connectivity information inside ASE Atoms objects, molify introduces a custom attribute ``ase.Atoms.info['connectivity']``: a list of tuples ``(atom_index_1, atom_index_2, bond_order)`` where indices are 0-based integers and ``bond_order`` can be a float (1.0, 2.0, 3.0, 1.5 for aromatic) or None for unknown bond orders.
+
+When connectivity information is not available:
+
+- **ase2networkx**: Uses covalent radii to estimate bonds between atoms.
+- **networkx2rdkit**: Uses ``rdkit.Chem.rdDetermineBonds.DetermineBonds`` to determine bond orders from 3D coordinates.
+
+
+Quick Start
+-----------
+
+**Basic Conversions**
+
+We can convert any RDKit molecule to an ASE Atoms object:
+
+.. code-block:: python
+
+   import molify
+   from rdkit import Chem
+
+   etoh = Chem.MolFromSmiles("CCO")
+   etoh = Chem.AddHs(etoh)
+
+   atoms = molify.rdkit2ase(etoh)
+   
+
+Likewise, we can convert the ASE Atoms object back to an RDKit molecule:
+
+.. code-block:: python
+
+   import molify
+   import ase.build
+
+   nh3 = ase.build.molecule("NH3")
+   mol = molify.ase2rdkit(nh3)
+
+**NetworkX Conversions**
+
+The networkx.Graph representation can be obtained from either RDKit or ASE objects:
+
+.. code-block:: python
+
+   import molify
+   import ase.build
+   from rdkit import Chem
+
+   etoh_mol = Chem.MolFromSmiles("CCO")
+   nh3_atoms = ase.build.molecule("NH3")
+
[…]
