+++ pyproject.toml
@@ -28,6 +28,7 @@ docs = [
     "nbsphinx>=0.9.7",
     "sphinx>=8.1.3",
     "sphinx-copybutton>=0.5.2",
+    "sphinxcontrib-mermaid>=1.0.0",
 ]
 
 [project.optional-dependencies]
