+++ molify/__init__.py
@@ -9,10 +9,9 @@ from molify.rdkit2x import rdkit2ase, rdkit2networkx
 from molify.smiles2x import smiles2atoms, smiles2conformers
 from molify.substructure import (
     get_substructures,
+    group_matches_by_fragment,
     iter_fragments,
     match_substructure,
-    select_atoms_flat_unique,
-    select_atoms_grouped,
     visualize_selected_molecules,
 )
 from molify.utils import draw_molecular_graph, unwrap_structures
@@ -26,12 +25,11 @@ __all__ = [
     "smiles2conformers",
     "compress",
     "match_substructure",
+    "group_matches_by_fragment",
     "get_substructures",
     "iter_fragments",
-    "select_atoms_flat_unique",
     "visualize_selected_molecules",
     "unwrap_structures",
-    "select_atoms_grouped",
     "draw_molecular_graph",
     #
     "ase2networkx",
