+++ tests/test_substructure.py
@@ -6,72 +6,33 @@ import molify
 
 def test_match_substructure():
     atoms = molify.smiles2atoms("CC(=O)O")
+    mol = molify.ase2rdkit(atoms, suggestions=["CC(=O)O"])
 
-    # match CH3 fragment using smarts
-    match = molify.match_substructure(atoms, smarts="[C]([H])([H])[H]")
+    # match CH3 fragment (carbon with 3 hydrogens)
+    match = molify.match_substructure(mol, "[C;H3]", hydrogens="include")
     assert match == ((0, 4, 5, 6),)
     assert atoms[match[0]].get_chemical_symbols() == ["C", "H", "H", "H"]
 
-    # now match using a ase.Atoms object
-
-    ref = molify.smiles2atoms("[C]([H])([H])[H]")
-    match = molify.match_substructure(atoms, fragment=ref)
-    assert match == ((0, 4, 5, 6),)
-
-    # now match using a Chem.Mol object
-    ref_mol = Chem.MolFromSmarts("[C]([H])([H])[H]")
-    match = molify.match_substructure(atoms, mol=ref_mol)
-    assert match == ((0, 4, 5, 6),)
-
-    # check everything else raises TypeError
-    with pytest.raises(TypeError):
-        molify.match_substructure(atoms, 42)  # type: ignore[arg-type]
-
 
 @pytest.mark.parametrize("packmol", ["packmol.jl"])
-def test_match_substructur_box(packmol):
+def test_match_substructure_box(packmol):
     atoms = molify.smiles2conformers("CC(=O)O", 1)
     box = molify.pack([atoms], counts=[3], packmol=packmol, density=0.5)
+    mol = molify.ase2rdkit(box, suggestions=["CC(=O)O"])
 
-    # match CH3 fragment using smarts
-    match = molify.match_substructure(box, smarts="[C]([H])([H])[H]")
+    # match CH3 fragment (carbon with 3 hydrogens)
+    match = molify.match_substructure(mol, "[C;H3]", hydrogens="include")
     assert match == ((0, 4, 5, 6), (8, 12, 13, 14), (16, 20, 21, 22))
     for m in match:
         assert box[m].get_chemical_symbols() == ["C", "H", "H", "H"]
-
-    # now match using a ase.Atoms object
-    ref = molify.smiles2atoms("[C]([H])([H])[H]")
-    match = molify.match_substructure(box, fragment=ref)
-    assert match == ((0, 4, 5, 6), (8, 12, 13, 14), (16, 20, 21, 22))
-
-    # now match using a Chem.Mol object
-    ref_mol = Chem.MolFromSmarts("[C]([H])([H])[H]")
-    match = molify.match_substructure(box, mol=ref_mol)
-    assert match == ((0, 4, 5, 6), (8, 12, 13, 14), (16, 20, 21, 22))
-
-    # check everything else raises TypeError
-    with pytest.raises(TypeError):
-        molify.match_substructure(box, 42)
 
 
 def test_get_substructure():
     atoms = molify.smiles2atoms("C(C(CO[N+](=O)[O-])O[N+](=O)[O-])O[N+](=O)[O-]")
     # match NO3 group using smarts
-    frames = molify.get_substructures(atoms, smarts="[N+](=O)[O-]")
-    assert len(frames) == 3
-    for frame in frames:
-        assert frame.get_chemical_symbols() == ["N", "O", "O"]
-
-    # match using a ase.Atoms object
-    ref = molify.smiles2atoms("[N+](=O)[O-]")
-    frames = molify.get_substructures(atoms, fragment=ref)
-    assert len(frames) == 3
-    for frame in frames:
[…]
