+++ molify/substructure.py
@@ -11,100 +11,271 @@ from molify.utils import find_connected_components
 
 
 def match_substructure(
-    atoms: ase.Atoms,
-    smiles: str | None = None,
-    smarts: str | None = None,
-    mol: Chem.Mol | None = None,
-    fragment: ase.Atoms | None = None,
-    **kwargs,
-) -> tuple[tuple[int, ...]]:
-    """
-    Find all matches of a substructure pattern in a given ASE Atoms object.
+    mol: Chem.Mol,
+    smarts_or_smiles: str,
+    hydrogens: tp.Literal["include", "exclude", "isolated"] = "exclude",
+    mapped_only: bool = False,
+) -> tuple[tuple[int, ...], ...]:
+    """Find all matches of a substructure pattern in an RDKit molecule.
+
+    This function performs substructure matching with advanced features including
+    hydrogen handling and atom mapping support. All matches are returned, including
+    matches across multiple disconnected fragments.
 
     Parameters
     ----------
-    atoms : ase.Atoms
-        The molecule or structure in which to search for substructure matches.
-    smiles : str, optional
-        A SMILES string representing the substructure pattern to match.
-    smarts : str, optional
-        A SMARTS string representing the substructure pattern to match.
-    mol : Chem.Mol, optional
-        An RDKit Mol object representing the substructure pattern to match.
-    fragment : ase.Atoms, optional
-        An ASE Atoms object representing the substructure pattern to match.
-        If provided, it will be converted to an RDKit Mol object for matching.
-    **kwargs
-        Additional keyword arguments passed to `ase2rdkit`.
+    mol : Chem.Mol
+        RDKit molecule to search. Use ``molify.ase2rdkit(atoms, suggestions=[...])``
+        to convert from ASE Atoms with explicit control over bond detection.
+    smarts_or_smiles : str
+        SMARTS pattern or mapped SMILES pattern (e.g., "[C:1][C:2]O").
+        If atom maps are present and ``mapped_only=True``, only mapped atoms
+        are returned in map number order.
+    hydrogens : {'exclude', 'include', 'isolated'}, default='exclude'
+        Controls hydrogen atom inclusion in matches:
+
+        - 'exclude': Return only heavy atoms (default)
+        - 'include': Return heavy atoms followed by their bonded hydrogens
+        - 'isolated': Return only hydrogens bonded to matched heavy atoms
+    mapped_only : bool, default=False
+        If True and pattern contains atom maps (e.g., [C:1]), return only the
+        mapped atoms ordered by map number. If False, return all matched atoms.
 
     Returns
     -------
-    tuple of tuple of int
-        A tuple of atom index tuples, each corresponding to one match of the pattern.
+    tuple[tuple[int, ...], ...]
+        Tuple of matches, where each match is a tuple of atom indices.
+        For patterns with multiple matches, each match is returned as a separate tuple.
 
+    Raises
+    ------
+    ValueError
+        If the SMARTS/SMILES pattern is invalid or if atom map labels are used
+        multiple times within the same pattern.
+
+    Examples
+    --------
+    >>> from molify import smiles2atoms, ase2rdkit, match_substructure
+    >>>
+    >>> # Explicit conversion with suggestions for better bond detection
+    >>> atoms = smiles2atoms("CCO")
+    >>> mol = ase2rdkit(atoms, suggestions=["CCO"])
+    >>>
+    >>> # Find all carbon atoms
+    >>> match_substructure(mol, "[#6]")
+    ((0,), (1,))
+    >>>
+    >>> # Find C-O bond with hydrogens included
[…]
