### 2025-10-27 d5b6cfe
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.12.12 → v0.14.2](https://github.com/astral-sh/ruff-pre-commit/compare/v0.12.12...v0.14.2)
- [github.com/executablebooks/mdformat: 0.7.22 → 1.0.0](https://github.com/executablebooks/mdformat/compare/0.7.22...1.0.0)
