### 2025-11-05 c8318e3
add auto-detection
