+++ molify/utils.py
@@ -229,78 +229,6 @@ def suggestions2networkx(smiles: list[str]) -> list[nx.Graph]:
     return [rdkit2networkx(mol) for mol in mols]
 
 
-def get_packmol_julia_version() -> str:
-    """Get the Packmol version when using Julia.
-
-    Raises
-    ------
-    RuntimeError
-        If the Packmol version cannot be retrieved.
-    """
-    try:
-        result = subprocess.run(
-            ["julia", "-e", 'import Pkg; Pkg.status("Packmol")'],
-            capture_output=True,
-            text=True,
-            check=True,
-        )
-        for line in result.stdout.splitlines():
-            if "Packmol" in line:
-                parts = line.split()
-                if len(parts) >= 3:
-                    return parts[2]
-        raise RuntimeError(
-            "Failed to get Packmol version via Julia. Please verify that you can"
-            ' import packmol via `julia -e "import Pkg; Pkg.status("Packmol")"`'
-        )
-
-    except subprocess.CalledProcessError as e:
-        raise RuntimeError(
-            "Failed to get Packmol version via Julia. Please verify that you can"
-            ' import packmol via `julia -e "import Pkg; Pkg.status("Packmol")"`'
-        ) from e
-
-
-def find_packmol_executable() -> str:
-    """Auto-detect available packmol installation.
-
-    Searches for packmol installations in the following order:
-    1. packmol.jl (Julia wrapper) - preferred for cross-platform compatibility
-    2. packmol binary in system PATH
-
-    Returns
-    -------
-    str
-        Either "packmol.jl" if Julia + Packmol package is found,
-        or the path to the packmol binary if found in system PATH.
-
-    Raises
-    ------
-    RuntimeError
-        If no packmol installation is found, with instructions for installation.
-
-    Examples
-    --------
-    >>> executable = find_packmol_executable()
-    >>> print(f"Found packmol: {executable}")
-    """
-    # Try Julia version first (preferred)
-    try:
-        get_packmol_julia_version()  # This will raise if not available
-        return "packmol.jl"
-    except RuntimeError:
-        pass  # Julia version not available, try native binary
-
-    # Try native packmol binary in PATH
-    packmol_path = shutil.which("packmol")
-    if packmol_path is not None:
-        return "packmol"
-
-    # Neither found - raise helpful error
-    raise RuntimeError(
-        "Could not find packmol installation.\n\n"
-        "See https://m3g.github.io/packmol/download.shtml."
-    )
 
 
 def draw_molecular_graph(
