### 2025-11-14 527c4f5
`build_data['pure_python'] = False`
