### 2025-11-14 438c727
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
