### 2025-11-14 6c649ea
brew install gcc before build
