### 2025-11-14 f83242c
search for gfortran
