+++ .github/workflows/build-wheels.yaml
@@ -26,7 +26,7 @@ jobs:
 
       - name: Build wheels
         uses: pypa/cibuildwheel@v3.3.0
-        env:
+        # env:
           # Skip all Windows builds for now (packmol needs mingw-w64/gfortran setup)
           # CIBW_SKIP: "*-win*"
           # On Linux x86_64 runners, build native + ARM via QEMU
@@ -36,7 +36,7 @@ jobs:
           # CIBW_ARCHS_MACOS: native
           # Install gfortran before building on each platform
           # CIBW_BEFORE_ALL_LINUX: yum install -y gcc-gfortran || apt-get update && apt-get install -y gfortran
-          CIBW_BEFORE_ALL_MACOS: brew install gcc
+          # CIBW_BEFORE_ALL_MACOS: brew install gcc
 
       - name: Upload wheels
         uses: actions/upload-artifact@v4
