### 2025-11-14 09a1d86
Handle packmol.exe on Windows platforms
- Check for both 'packmol' and 'packmol.exe' after compilation
- Windows builds produce packmol.exe, Unix/Linux/macOS produce packmol
- Always install as 'packmol' (no extension) for cross-platform consistency

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
