### 2025-11-14 26c3ef6
remove `packmol.jl`
