+++ tests/conftest.py
@@ -9,7 +9,7 @@ def ethanol_water():
     ethanol = molify.smiles2conformers("CCO", numConfs=100)
     water = molify.smiles2conformers("O", numConfs=100)
     box = molify.pack(
-        [ethanol, water], counts=[2, 2], density=700, packmol="packmol.jl"
+        [ethanol, water], counts=[2, 2], density=700
     )
     return box.copy()
 
@@ -24,6 +24,6 @@ def alanine_dipeptide() -> ase.Atoms:
 def alanine_dipeptide_box(alanine_dipeptide) -> ase.Atoms:
     """Box of alanine dipeptide molecules using molify.pack"""
     box = molify.pack(
-        [[alanine_dipeptide]], counts=[3], density=500, packmol="packmol.jl"
+        [[alanine_dipeptide]], counts=[3], density=500
     )
     return box.copy()
