### 2025-11-14 d1d4c61
Fix macOS deployment target and Windows Unicode error
- Set MACOSX_DEPLOYMENT_TARGET=14.0 for macOS wheel compatibility
- Remove Unicode checkmark (✓) that causes UnicodeEncodeError on Windows
- Add CIBW_BEFORE_ALL_MACOS to install gcc before building

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
