### 2025-11-14 3f6917e
Add PyPI publishing and improve workflow
- Remove gcc installation (pre-installed on runners)
- Rename build_sdist → make_sdist with fetch-depth: 0
- Add upload_all job for PyPI publishing on releases
- Use trusted publishing with pypa/gh-action-pypi-publish
- Download all cibw-* artifacts before publishing

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
