### 2025-11-14 8cd78eb
run pytest on all versions
