+++ .github/workflows/publish.yaml
@@ -1,19 +1,0 @@
-name: Release
-on:
-  release:
-    types:
-      - created
-
-jobs:
-  publish:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-      - name: Install uv
-        uses: astral-sh/setup-uv@v5
-      - name: Publish
-        env:
-          PYPI_TOKEN: ${{ secrets.PYPI_API_TOKEN }}
-        run: |
-          uv build
-          uv publish --token $PYPI_TOKEN
