+++ .github/workflows/pytest.yaml
@@ -21,13 +21,12 @@ jobs:
           - "3.12"
           - "3.11"
           - "3.10"
-        os:
-          - ubuntu-latest
-        packmol-version:
-          - "20.15.1"
+        os: [ubuntu-latest, ubuntu-24.04-arm, windows-latest, windows-11-arm, macos-15-intel, macos-14]
 
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v5
+        with:
+          submodules: recursive
       - name: Install uv and set the python version
         uses: astral-sh/setup-uv@v5
         with:
@@ -36,9 +35,6 @@ jobs:
         run: uv sync --all-extras --dev
       - name: Pytest
         run: |
-          # add packmol to path
-          export PATH=$PATH:$(pwd)/packmol-${{ matrix.packmol-version}}
-
           uv run python --version
           uv run pytest --cov --junitxml=junit.xml -o junit_family=legacy
       - name: Upload coverage to Codecov
