+++ .github/workflows/build-wheels.yaml
@@ -18,12 +18,6 @@ jobs:
         with:
           submodules: recursive  # Critical: fetch packmol submodule
 
-      # - name: Set up QEMU (for Linux x86_64 cross-compilation to ARM)
-      #   if: runner.os == 'Linux' && runner.arch == 'X64'
-      #   uses: docker/setup-qemu-action@v3
-      #   with:
-      #     platforms: arm64
-
       - name: Build wheels
         uses: pypa/cibuildwheel@v3.3.0
         env:
