### 2025-11-14 86c58c7
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
