+++ tests/conftest.py
@@ -8,9 +8,7 @@ import molify
 def ethanol_water():
     ethanol = molify.smiles2conformers("CCO", numConfs=100)
     water = molify.smiles2conformers("O", numConfs=100)
-    box = molify.pack(
-        [ethanol, water], counts=[2, 2], density=700
-    )
+    box = molify.pack([ethanol, water], counts=[2, 2], density=700)
     return box.copy()
 
 
@@ -23,7 +21,5 @@ def alanine_dipeptide() -> ase.Atoms:
 @pytest.fixture(scope="session")
 def alanine_dipeptide_box(alanine_dipeptide) -> ase.Atoms:
     """Box of alanine dipeptide molecules using molify.pack"""
-    box = molify.pack(
-        [[alanine_dipeptide]], counts=[3], density=500
-    )
+    box = molify.pack([[alanine_dipeptide]], counts=[3], density=500)
     return box.copy()
