### 2025-11-14 f6a9a56
update build targets
