+++ .github/workflows/pytest.yaml
@@ -21,7 +21,7 @@ jobs:
           - "3.12"
           - "3.11"
           - "3.10"
-        os: [ubuntu-latest, ubuntu-24.04-arm, windows-latest, windows-11-arm, macos-15-intel, macos-14]
+        os: [ubuntu-latest, ubuntu-24.04-arm, windows-latest, windows-11-arm, macos-15-intel, macos-latest]
 
     steps:
       - uses: actions/checkout@v5
