### 2025-11-14 fcb6389
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
