### 2025-11-14 d10fe61
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
