+++ .github/workflows/build-wheels.yaml
@@ -20,7 +20,7 @@ jobs:
         uses: pypa/cibuildwheel@v3.3.0
         env:
           # Set macOS deployment target dynamically based on runner
-          CIBW_ENVIRONMENT_MACOS: MACOSX_DEPLOYMENT_TARGET=${{ matrix.os == 'macos-15-intel' && '15.0' || '14.0' }}
+          CIBW_ENVIRONMENT_MACOS: MACOSX_DEPLOYMENT_TARGET=15.0
 
       - name: Upload wheels
         uses: actions/upload-artifact@v4
