### 2025-11-14 aeb14ce
remove `windows-11-arm` tests
