### 2025-11-15 93c9cf6
build wheels (#83)
* build wheels

* `build_data['pure_python'] = False`

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* brew install gcc before build

* search for gfortran

* Handle packmol.exe on Windows platforms

- Check for both 'packmol' and 'packmol.exe' after compilation
- Windows builds produce packmol.exe, Unix/Linux/macOS produce packmol
- Always install as 'packmol' (no extension) for cross-platform consistency

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>

[…]
