### 2025-11-15 072edc5
update docs / bump version
