### 2025-11-15 c8c0df9
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
