+++ .github/workflows/build-wheels.yaml
@@ -1,6 +1,14 @@
-name: Build
+name: Build and upload to PyPI
 
-on: [push, pull_request]
+on:
+  workflow_dispatch:
+  pull_request:
+  push:
+    branches:
+      - main
+  release:
+    types:
+      - published
 
 jobs:
   build_wheels:
@@ -28,8 +36,8 @@ jobs:
           name: cibw-wheels-${{ matrix.os }}-${{ strategy.job-index }}
           path: ./wheelhouse/*.whl
 
-  make_sdist:
-    name: Make SDist
+  build_sdist:
+    name: Build source distribution
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v5
@@ -49,8 +57,8 @@ jobs:
           name: cibw-sdist
           path: dist/*.tar.gz
 
-  upload_all:
-    needs: [build_wheels, make_sdist]
+  upload_pypi:
+    needs: [build_wheels, build_sdist]
     environment: pypi
     permissions:
       id-token: write
@@ -58,7 +66,7 @@ jobs:
     if: github.event_name == 'release' && github.event.action == 'published'
     steps:
       - name: Download all artifacts
-        uses: actions/download-artifact@v4
+        uses: actions/download-artifact@v5
         with:
           pattern: cibw-*
           path: dist
