### 2025-11-15 82600fe
update documentation (#80)
* update documentation

* update docs

* update docs

* lint / format

* update docs

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* add auto-detection

* build wheels (#83)

* build wheels

* `build_data['pure_python'] = False`

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* brew install gcc before build

* search for gfortran

* Handle packmol.exe on Windows platforms

- Check for both 'packmol' and 'packmol.exe' after compilation
- Windows builds
[…]
