+++ tests/test_com.py
@@ -104,10 +104,7 @@ def test_com_with_packed_system(shift):
     density = 900  # kg/m^3
 
     packed_system = molify.pack(
-        [water_template, ethanol_template],
-        [num_water, num_ethanol],
-        density,
-        packmol="packmol.jl",
+        [water_template, ethanol_template], [num_water, num_ethanol], density
     )
     packed_system.positions += np.array([shift, shift, shift])
     packed_system.wrap()
@@ -142,7 +139,6 @@ def test_get_centers_of_mass_species():
         [a, b, c, d],
         [1, 2, 3, 4],
         density=786,
-        packmol="packmol.jl",
     )
 
     com = molify.get_centers_of_mass(box)

[…]
