### 2025-11-15 zincware/molify#69 @PythonFZ
I've removed everything related to `packmol.jl`. The `molify` package now includes the `packmol` binary on most platforms, making a dedicated packmol installation obsolete.

https://github.com/zincware/rdkit2ase/blob/82600fe0524fd43f03e0983fd19f425b798f1769/.github/workflows/build-wheels.yaml#L20

If this fails, one can still set the `packmol=/path/to/packmol`. I will make a full release (not pre-release) after reviewing the other comments. with https://pypi.org/project/molify/0.0.1a2/ all features should work out of the box.
