### 2025-11-15 24f9f17
[pre-commit.ci] pre-commit autoupdate (#76)
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.12.12 → v0.14.4](https://github.com/astral-sh/ruff-pre-commit/compare/v0.12.12...v0.14.4)
- [github.com/executablebooks/mdformat: 0.7.22 → 1.0.0](https://github.com/executablebooks/mdformat/compare/0.7.22...1.0.0)

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
