+++ .pre-commit-config.yaml
@@ -26,7 +26,7 @@ repos:
         additional_dependencies: ["tomli"]
   - repo: https://github.com/astral-sh/ruff-pre-commit
     # Ruff version.
-    rev: v0.12.12
+    rev: v0.14.4
     hooks:
       # Run the linter.
       - id: ruff
@@ -34,7 +34,7 @@ repos:
       # Run the formatter.
       - id: ruff-format
   - repo: https://github.com/executablebooks/mdformat
-    rev: 0.7.22
+    rev: 1.0.0
     hooks:
     - id: mdformat
       args: ["--wrap=80"]
