+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'rdkit2ase: Molecular Structure Generation and Manipulation for Machine-Learned Interatomic Potentials'
+title: 'molify: Molecular Structure Interface'
 tags:
   - Python
   - cheminformatics
@@ -23,24 +23,27 @@ bibliography: bibliography.bib
 The increasing prevalence of Machine-Learned Interatomic Potentials (MLIPs) has shifted requirements for setting up atomistic simulations.
 Unlike classical force fields, MLIPs primarily require atomic positions and species, thereby removing the need for predefined topology files used for classical force fields in molecular dynamics software like GROMACS[@abrahamGROMACSHighPerformance2015], LAMMPS[@LAMMPS], ESPResSo[@weikESPResSo40Extensible2019] or OpenMM [@eastmanOpenMM8Molecular2024].
 Consequently, the Atomic Simulation Environment (ASE) [@larsenAtomicSimulationEnvironment2017] has become a popular Python toolkit for handling atomic structures and interfacing with MLIPs, particularly within the materials science and soft matter communities, because it originates from _ab initio_ simulations, which share the same setup as MLIP-driven studies.
+In contrast to _ab initio_, MLIPs are much faster and can run on much larger systems, making high throughput simulations of more complex systems feasible and increasing the need for efficient initial structure generation.
 
 Concurrently, RDKit [@landrumRdkitRdkit2023_03_22023] offers extensive functionality for cheminformatics and manipulating chemical structures.
 However, standard RDKit workflows are not designed for MLIP-driven simulation, while typical ASE-MLIP workflows may lack rich, explicit chemical information such as bond orders or molecular identities, as well as generating different conformations or searching substructures.
 
-The rdkit2ase package bridges this gap, providing an interface between RDKit's chemical structure generation and cheminformatics capabilities and ASE's handling of 3D atomic structures.
-Furthermore, rdkit2ase integrates with PACKMOL [@martinezPACKMOLPackageBuilding2009] to facilitate the creation of complex, periodic simulation cells with diverse chemical compositions, all while preserving crucial chemical connectivity information.
+The `molify` package bridges this gap, providing an interface between RDKit's chemical structure generation and cheminformatics capabilities and ASE's handling of 3D atomic structures.
+Furthermore, `molify` integrates with PACKMOL [@martinezPACKMOLPackageBuilding2009] to facilitate the creation of complex, periodic simulation cells with diverse chemical compositions, all while preserving crucial chemical connectivity information.
+In addition, `molify` simplifies the representation of molecular structures as graphs using NetworkX [@hagbergExploringNetworkStructure2008], e.g. enabling traversing or comparing molecular graphs. 
 Lastly, the combination of these packages enables selection and manipulation of atomistic structures based on chemical knowledge rather than manual index handling.
 
 
 # Statement of need
-rdkit2ase serves as a vital link between RDKit, ASE, and PACKMOL.
+`molify` serves as a vital link between RDKit, ASE, NetworkX, and PACKMOL.
 While its core function is to interface these tools, it thereby unlocks new capabilities and significantly reduces the manual coding and data wrangling typically required for preparing and analyzing molecular simulations.
+For example, ASE has no tools for handling topological information such as bonds or molecular identities, whilst rdkit does not natively interface with MLIPs.
 
[…]
