+++ paper/bibliography.bib
@@ -210,3 +210,17 @@
   year      = {2008},
   editor    = {Ga\"el Varoquaux and Travis Vaught and Jarrod Millman}
 }
+@article{oboyleOpenBabelOpen2011,
+  title = {Open {{Babel}}: {{An}} Open Chemical Toolbox},
+  shorttitle = {Open {{Babel}}},
+  author = {O'Boyle, Noel M. and Banck, Michael and James, Craig A. and Morley, Chris and Vandermeersch, Tim and Hutchison, Geoffrey R.},
+  year = 2011,
+  month = oct,
+  journal = {Journal of Cheminformatics},
+  volume = {3},
+  number = {1},
+  pages = {33},
+  issn = {1758-2946},
+  doi = {10.1186/1758-2946-3-33},
+  urldate = {2025-11-15},
+}
