### 2025-11-17 f09107d
update manuscript
