### 2025-11-17 53ec268
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
