+++ README.md
@@ -1,14 +1,14 @@
 [![zincware](https://img.shields.io/badge/Powered%20by-zincware-darkcyan)](https://github.com/zincware)
-[![codecov](https://codecov.io/gh/zincware/rdkit2ase/graph/badge.svg?token=Q0VIN03185)](https://codecov.io/gh/zincware/rdkit2ase)
+[![codecov](https://codecov.io/gh/zincware/molify/graph/badge.svg?token=Q0VIN03185)](https://codecov.io/gh/zincware/molify)
 [![PyPI version](https://badge.fury.io/py/molify.svg)](https://badge.fury.io/py/molify)
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.15423477.svg)](https://doi.org/10.5281/zenodo.15423477)
-[![Docs](https://github.com/zincware/rdkit2ase/actions/workflows/pages.yaml/badge.svg)](https://zincware.github.io/rdkit2ase/)
+[![Docs](https://github.com/zincware/molify/actions/workflows/pages.yaml/badge.svg)](https://zincware.github.io/molify/)
 [![status](https://joss.theoj.org/papers/4c1ec2a0b66ed6baaf3fe93066696d5f/status.svg)](https://joss.theoj.org/papers/4c1ec2a0b66ed6baaf3fe93066696d5f)
 
 # molify - Molecular Structure Interface with RDKit, ASE, Packmol and NetworkX
 
 Installation via `pip install molify`. For more information please visit the
-[documentation](https://zincware.github.io/rdkit2ase/).
+[documentation](https://zincware.github.io/molify/).
 
 A common use case is to create 3D structures from SMILES strings. This can be
 achieved using the `molify.rdkit2ase` function.
