### 2025-11-17 df9c6e8
use packmol from PyPI
