+++ pyproject.toml
@@ -12,7 +12,7 @@ dependencies = [
     "ase>=3.25.0",
     "matplotlib>=3.10.1",
     "networkx>=3.4.2",
-    "packmol>=0.1.0",
+    "packmol>=21.1.2",
     "rdkit>=2024",
 ]
 
