### 2025-11-17 f61a221
remove entry point
