### 2025-11-17 4966bb7
use packmol from PyPI (#87)
* use packmol from PyPI

* bump packmol version

* remove entry point
