### 2025-11-25 #8829 @lucydot
@csadorf - I can see a few outstanding items on your checklist - could you give an update on this?
