### 2025-11-25 66da46f
Add tests for graph modification workflow (ASE -> networkx -> modify -> ASE/rdkit)
Co-authored-by: PythonFZ <46721498+PythonFZ@users.noreply.github.com>
