+++ tests/integration/test_networkx2x.py
@@ -313,3 +313,294 @@ def test_networkx2rdkit_error_on_failed_determination():
     # This should raise an error (from underlying bond determination)
     with pytest.raises(ValueError, match="Failed to determine bonds"):
         molify.networkx2rdkit(graph, suggestions=[])
+
+
+# ============================================================================
+# Tests for graph modification workflows (remove/add nodes)
+# These tests verify the workflow: ASE -> networkx -> modify -> ASE/rdkit
+# ============================================================================
+
+
+def test_modify_graph_remove_nodes_to_ase():
+    """Test ASE -> networkx -> remove nodes -> ASE workflow.
+
+    Verifies that after removing nodes from a networkx graph,
+    the resulting ASE atoms object has correct structure.
+    """
+    # Create ethanol molecule (C-C-O with hydrogens)
+    atoms = molify.smiles2atoms("CCO")  # Ethanol
+    graph = molify.ase2networkx(atoms)
+
+    original_num_nodes = graph.number_of_nodes()
+
+    # Identify hydrogen nodes to remove (keep only heavy atoms)
+    heavy_atom_nodes = [
+        n for n, d in graph.nodes(data=True) if d["atomic_number"] != 1
+    ]
+
+    # Create modified graph by removing hydrogen nodes
+    modified_graph = graph.subgraph(heavy_atom_nodes).copy()
+
+    # Verify the modified graph has fewer nodes
+    assert modified_graph.number_of_nodes() < original_num_nodes
+    assert modified_graph.number_of_nodes() == len(heavy_atom_nodes)
+
+    # Convert back to ASE
+    new_atoms = molify.networkx2ase(modified_graph)
+
+    # Verify the new atoms object
+    assert len(new_atoms) == len(heavy_atom_nodes)
+    assert all(num != 1 for num in new_atoms.get_atomic_numbers())
+
+    # Verify connectivity is valid
+    connectivity = new_atoms.info["connectivity"]
+    for i, j, bond_order in connectivity:
+        assert 0 <= i < len(new_atoms)
+        assert 0 <= j < len(new_atoms)
+
+
+def test_modify_graph_remove_nodes_to_rdkit():
+    """Test ASE -> networkx -> remove nodes -> rdkit workflow.
+
+    Verifies that after removing nodes from a networkx graph,
+    the graph can be converted to an RDKit molecule.
+
+    This test keeps bond orders intact when creating subgraph,
+    which is the typical use case when you know the molecular structure.
+    """
+    # Create ethanol molecule
+    atoms = molify.smiles2atoms("CCO")  # Ethanol
+    graph = molify.ase2networkx(atoms)
+
+    original_num_nodes = graph.number_of_nodes()
+
+    # Get the oxygen atom and its connected atoms (OH group with one H)
+    oxygen_nodes = [n for n, d in graph.nodes(data=True) if d["atomic_number"] == 8]
+    assert len(oxygen_nodes) == 1
+    oxygen_node = oxygen_nodes[0]
+
+    # Find the H attached to oxygen
+    oh_hydrogen = None
+    for neighbor in graph.neighbors(oxygen_node):
+        if graph.nodes[neighbor]["atomic_number"] == 1:
+            oh_hydrogen = neighbor
+            break
+
+    # Remove just the OH hydrogen - keeping bond orders from original connectivity
+    nodes_to_keep = [n for n in graph.nodes() if n != oh_hydrogen]
+    modified_graph = graph.subgraph(nodes_to_keep).copy()
+
[…]
