### 2025-11-25 1324242
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
