### 2025-11-25 2d0db47
Add tests for graph modification workflow (ASE → networkx → modify → ASE/rdkit) (#90)
* Initial plan

* Add tests for graph modification workflow (ASE -> networkx -> modify -> ASE/rdkit)

Co-authored-by: PythonFZ <46721498+PythonFZ@users.noreply.github.com>

* Improve test comments for clarity on smiles2atoms behavior

Co-authored-by: PythonFZ <46721498+PythonFZ@users.noreply.github.com>

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: copilot-swe-agent[bot] <198982749+Copilot@users.noreply.github.com>
Co-autho
[…]
