### 2025-11-25 #8829 @PythonFZ
I have removed the statement on the `packmol` binary and some smaller language changes. I added a very short section on converting from / to networkx. Here is the diff https://github.com/zincware/molify/pull/48/commits/76c4b20cc95fac11be3be704e41fcfdbba834a01
