### 2025-11-26 d74bfec
Add DeepWiki badge to README (#91)
