+++ README.md
@@ -4,6 +4,7 @@
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.15423477.svg)](https://doi.org/10.5281/zenodo.15423477)
 [![Docs](https://github.com/zincware/molify/actions/workflows/pages.yaml/badge.svg)](https://zincware.github.io/molify/)
 [![status](https://joss.theoj.org/papers/4c1ec2a0b66ed6baaf3fe93066696d5f/status.svg)](https://joss.theoj.org/papers/4c1ec2a0b66ed6baaf3fe93066696d5f)
+[![Ask DeepWiki](https://deepwiki.com/badge.svg)](https://deepwiki.com/zincware/molify)
 
 # molify - Molecular Structure Interface with RDKit, ASE, Packmol and NetworkX
 
