### 2025-11-26 42f9d4c
[pre-commit.ci] pre-commit autoupdate (#88)
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.14.4 → v0.14.6](https://github.com/astral-sh/ruff-pre-commit/compare/v0.14.4...v0.14.6)

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
