### 2025-12-02 #8829 @lucydot
@PythonFZ  - we are ready to move onto the final editorial checks 🎈 . Please let me know when you have worked through the author checklist below -
