### 2025-12-03 2056d87
add CITATION.cff (#96)
