### 2025-12-09 324e9ce
Editorial Comments (#98)
* JOSS manuscript draft

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* lint

* fix `mdformat` removing important parts of the file

* draft

* update manuscript

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* update manuscript

* editorial comments

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
