+++ paper/bibliography.bib
@@ -202,25 +202,24 @@
   urldate   = {2025-05-24}
 }
 @inproceedings{hagbergExploringNetworkStructure2008,
-  author    = {Aric A. Hagberg and Daniel A. Schult and Pieter J. Swart},
-  title     = {Exploring Network Structure, Dynamics, and Function using NetworkX},
-  booktitle = {Proceedings of the 7th Python in Science Conference},
-  pages     = {11 - 15},
-  address   = {Pasadena, CA USA},
-  year      = {2008},
-  editor    = {Ga\"el Varoquaux and Travis Vaught and Jarrod Millman}
+  author  = {Aric A. Hagberg and Daniel A. Schult and Pieter J. Swart},
+  year    = {2008},
+  month   = {06},
+  title   = {Exploring Network Structure, Dynamics, and Function Using NetworkX},
+  journal = {Proceedings of the 7th Python in Science Conference},
+  doi     = {10.25080/TCWV9851}
 }
 @article{oboyleOpenBabelOpen2011,
-  title = {Open {{Babel}}: {{An}} Open Chemical Toolbox},
+  title      = {Open {{Babel}}: {{An}} Open Chemical Toolbox},
   shorttitle = {Open {{Babel}}},
-  author = {O'Boyle, Noel M. and Banck, Michael and James, Craig A. and Morley, Chris and Vandermeersch, Tim and Hutchison, Geoffrey R.},
-  year = 2011,
-  month = oct,
-  journal = {Journal of Cheminformatics},
-  volume = {3},
-  number = {1},
-  pages = {33},
-  issn = {1758-2946},
-  doi = {10.1186/1758-2946-3-33},
-  urldate = {2025-11-15},
+  author     = {O'Boyle, Noel M. and Banck, Michael and James, Craig A. and Morley, Chris and Vandermeersch, Tim and Hutchison, Geoffrey R.},
+  year       = 2011,
+  month      = oct,
+  journal    = {Journal of Cheminformatics},
+  volume     = {3},
+  number     = {1},
+  pages      = {33},
+  issn       = {1758-2946},
+  doi        = {10.1186/1758-2946-3-33},
+  urldate    = {2025-11-15}
 }
