### 2025-12-12 #8829 @Kevin-Mattheus-Moerman
- [ ] Please add a heading for the `References` after Acknowledgements.
