### 2025-12-12 38e3f41
add `# References`
