+++ paper/paper.md
@@ -137,3 +137,5 @@ Furthermore, `molify`'s integration with PACKMOL and NetworkX provides capabilit
 
 # Acknowledgements
 F.Z. acknowledges support by the Deutsche Forschungsgemeinschaft (DFG, German Research Foundation) in the framework of the priority program SPP 2363, "Utilization and Development of Machine Learning for Molecular Applications – Molecular Machine Learning" Project No. 497249646. Further funding through the DFG under Germany's Excellence Strategy – EXC 2075 – 390740016 and the Stuttgart Center for Simulation Science (SimTech) was provided.
+
+# References
