### 2026-01-26 fb495e5
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.14.7 → v0.14.14](https://github.com/astral-sh/ruff-pre-commit/compare/v0.14.7...v0.14.14)
